Algorithm Traits

Summary. Traits allow one to promise particular behaviour for an algorithm, such as: This algorithm supports per-observation weights, which must appear as the third argument of fit, or This algorithm's transform method predicts Real vectors.

Algorithm traits are functions whose first (and usually only) argument is an algorithm. In a new implementation, a single-argument trait is declared following this pattern:

LearnAPI.is_pure_julia(algorithm::MyAlgorithmType) = true

Important

The value of a trait must be the same for all algorithms of the same type, even if the types differ only in type parameters. There are exceptions for some traits, if is_wrapper(algorithm) = true for all instances algorithm of some type (composite algorithms). This requirement occasionally requires that an existing algorithm implementation be split into separate LearnAPI implementations (e.g., one for regression and another for classification).

The declaration above has the shorthand

@trait MyAlgorithmType is_pure_julia=true

Multiple traits can be declared like this:

@trait(
    MyAlgorithmType,
    is_pure_julia = true,
    pkg_name = "MyPackage",
)

Special two-argument traits

The two-argument version of LearnAPI.predict_output_scitype and LearnAPI.predict_output_scitype are the only overloadable traits with more than one argument. They cannot be declared using the @trait macro.

Trait summary

Overloadable traits are available for overloading by any new LearnAPI implementation. Derived traits are not, and should not be called by performance critical code

Overloadable traits

In the examples column of the table below, Table, Continuous, Sampleable are names owned by the package ScientificTypesBase.jl.

trait	fallback value	return value	example

LearnAPI.functions(algorithm)	()	implemented LearnAPI functions (traits excluded)	(:fit, :predict)

LearnAPI.preferred_kind_of_proxy(algorithm)	LearnAPI.None()	an instance tp of KindOfProxy for which an implementation of LearnAPI.predict(algorithm, tp, ...) is guaranteed.	LearnAPI.Distribution()

LearnAPI.position_of_target(algorithm)	0	¹ the positional index of the target in data in fit(..., data...; metadata) calls	2

LearnAPI.position_of_weights(algorithm)	0	¹ the positional index of per-observation weights in data in fit(..., data...; metadata)	3

LearnAPI.descriptors(algorithm)	()	lists one or more suggestive algorithm descriptors from LearnAPI.descriptors()	(:classifier, :probabilistic)

LearnAPI.is_pure_julia(algorithm)	false	is true if implementation is 100% Julia code	true

LearnAPI.pkg_name(algorithm)	"unknown"	name of package providing core code (may be different from package providing LearnAPI.jl implementation)	"DecisionTree"

LearnAPI.pkg_license(algorithm)	"unknown"	name of license of package providing core code	"MIT"

LearnAPI.doc_url(algorithm)	"unknown"	url providing documentation of the core code	"https://en.wikipedia.org/wiki/Decision_tree_learning"

LearnAPI.load_path(algorithm)	"unknown"	a string indicating where the struct for typeof(algorithm) is defined, beginning with name of package providing implementation	FastTrees.LearnAPI.DecisionTreeClassifier

LearnAPI.is_wrapper(algorithm)	false	is true if one or more properties (fields) of algorithm may be an algorithm	true

LearnAPI.human_name(algorithm)	type name with spaces	human name for the algorithm; should be a noun	"elastic net regressor"

LearnAPI.iteration_parameter(algorithm)	nothing	symbolic name of an iteration parameter	:epochs

LearnAPI.fit_keywords(algorithm)	()	tuple of symbols for keyword arguments accepted by fit (corresponding to metadata)	(:class_weights,)

LearnAPI.fit_scitype(algorithm)	Union{}	upper bound on scitype(data) in fit(algorithm, verbosity, data...)²	Tuple{Table(Continuous), AbstractVector{Continuous}}

LearnAPI.fit_observation_scitype(algorithm)	Union{}	upper bound on scitype(observation) for observation in data and data in fit(algorithm, verbosity, data...)²	Tuple{AbstractVector{Continuous}, Continuous}

LearnAPI.fit_type(algorithm)	Union{}	upper bound on type(data) in fit(algorithm, verbosity, data...)²	Tuple{AbstractMatrix{<:Real}, AbstractVector{<:Real}}

LearnAPI.fit_observation_type(algorithm)	Union{}	upper bound on type(observation) for observation in data and data in fit(algorithm, verbosity, data...)*	Tuple{AbstractVector{<:Real}, Real}

LearnAPI.predict_input_scitype(algorithm)	Union{}	upper bound on scitype(data) in predict(algorithm, fitted_params, data...)²	Table(Continuous)

LearnAPI.predict_output_scitype(algorithm, kind_of_proxy)	Any	upper bound on scitype(first(predict(algorithm, kind_of_proxy, ...)))	AbstractVector{Continuous}

LearnAPI.predict_input_type(algorithm)	Union{}	upper bound on typeof(data) in predict(algorithm, fitted_params, data...)²	AbstractMatrix{<:Real}

LearnAPI.predict_output_type(algorithm, kind_of_proxy)	Any	upper bound on typeof(first(predict(algorithm, kind_of_proxy, ...)))	AbstractVector{<:Real}

LearnAPI.transform_input_scitype(algorithm)	Union{}	upper bound on scitype(data) in transform(algorithm, fitted_params, data...)²	Table(Continuous)

LearnAPI.transform_output_scitype(algorithm)	Any	upper bound on scitype(first(transform(algorithm, ...)))	Table(Continuous)

LearnAPI.transform_input_type(algorithm)	Union{}	upper bound on typeof(data) in transform(algorithm, fitted_params, data...)²	AbstractMatrix{<:Real}}

LearnAPI.transform_output_type(algorithm)	Any	upper bound on typeof(first(transform(algorithm, ...)))	AbstractMatrix{<:Real}

¹ If the value is 0, then the variable in boldface type is not supported and not expected to appear in data. If length(data) is less than the trait value, then data is understood to exclude the variable, but note that fit can have multiple signatures of varying lengths, as in fit(algorithm, verbosity, X, y) and fit(algorithm, verbosity, X, y, w). A non-zero value is a promise that fit includes a signature of sufficient length to include the variable.

² Assuming no optional data interface is implemented. See docstring for the general case.

Derived Traits

The following convenience methods are provided but intended for overloading:

trait	return value	example

LearnAPI.name(algorithm)	algorithm type name as string	"PCA"

LearnAPI.is_algorithm(algorithm)	true if functions(algorithm) is not empty	true

LearnAPI.predict_output_scitype(algorithm)	dictionary of upper bounds on the scitype of predictions, keyed on subtypes of LearnAPI.KindOfProxy

LearnAPI.predict_output_type(algorithm)	dictionary of upper bounds on the type of predictions, keyed on subtypes of LearnAPI.KindOfProxy

Reference

LearnAPI.functions — Function

LearnAPI.functions(algorithm)

Return a tuple of symbols, such as (:fit, :predict), corresponding to LearnAPI methods specifically implemented for objects having the same type as algorithm. If non-empty, this also guarantees algorithm is an algorithm, in the LearnAPI sense. See the Reference section of the manual for details.

New implementations

Every LearnAPI method that is not a trait and which is specifically implemented for typeof(algorithm) must be included in the return value of this trait. Specifically, the return value is a tuple of symbols from this list: :fit, :update!, :ingest!, :predict, :transform, :inverse_transform, :features_importances, :training_labels, :training_losses, :training_scores. To regenerate this list, do LearnAPI.functions().

See also LearnAPI.Algorithm.

LearnAPI.preferred_kind_of_proxy — Function

LearnAPI.preferred_kind_of_proxy(algorithm)

Returns an instance of LearnAPI.KindOfProxy, unless LearnAPI.predict is not implemented for objects of type typeof(algorithm), in which case it returns nothing.

The returned target proxy is generally the one with the smallest computational cost, if more than one type is supported.

See also LearnAPI.predict, LearnAPI.KindOfProxy.

New implementations

Any algorithm implementing LearnAPI.predict must overload this trait.

The trait must return a lone instance T() for some concrete subtype T <: LearnAPI.KindOfProxy. List these with subtypes(LearnAPI.KindOfProxy) and subtypes(LearnAPI.IID).

Suppose, for example, we have the following implementation of a supervised learner returning only probablistic predictions:

LearnAPI.predict(algorithm::MyNewAlgorithmType, LearnAPI.Distribution(), Xnew) = ...

Then we can declare

@trait MyNewAlgorithmType  preferred_kind_of_proxy = LearnAPI.LiteralTarget()

which is shorthand for

LearnAPI.preferred_kind_of_proxy(::MyNewAlgorithmType) = LearnAPI.Distribution()

For more on target variables and target proxies, refer to the LearnAPI documentation.

LearnAPI.position_of_target — Function

LearnAPI.position_of_target(algorithm)

Return the expected position of the target variable within data in calls of the form LearnAPI.fit(algorithm, verbosity, data...).

If this number is 0, then no target is expected. If this number exceeds length(data), then data is understood to exclude the target variable.

LearnAPI.position_of_weights — Function

LearnAPI.position_of_weights(algorithm)

Return the expected position of per-observation weights within data in calls of the form LearnAPI.fit(algorithm, verbosity, data...).

If this number is 0, then no weights are expected. If this number exceeds length(data), then data is understood to exclude weights, which are assumed to be uniform.

LearnAPI.descriptors — Function

LearnAPI.descriptors(algorithm)

Lists one or more suggestive algorithm descriptors from this list: :regression, :classification, :clustering, :gradient_descent, :iterative_algorithms, :incremental_algorithms, :dimension_reduction, :encoders, :static_algorithms, :missing_value_imputation, :ensemble_algorithms, :wrappers, :time_series_forecasting, :time_series_classification, :survival_analysis, :distribution_fitters, :Bayesian_algorithms, :outlier_detection, :collaborative_filtering, :text_analysis, :audio_analysis, :natural_language_processing, :image_processing (do LearnAPI.descriptors() to reproduce).

Warning

The value of this trait guarantees no particular behavior. The trait is intended for informal classification purposes only.

New implementations

This trait should return a tuple of symbols, as in (:classifier, :probabilistic).

LearnAPI.is_pure_julia — Function

LearnAPI.is_pure_julia(algorithm)

Returns true if training algorithm requires evaluation of pure Julia code only.

New implementations

The fallback is false.

LearnAPI.pkg_name — Function

LearnAPI.pkg_name(algorithm)

Return the name of the package module which supplies the core training algorithm for algorithm. This is not necessarily the package providing the LearnAPI interface.

Returns "unknown" if the algorithm implementation has failed to overload the trait.

New implementations

Must return a string, as in "DecisionTree".

LearnAPI.pkg_license — Function

LearnAPI.pkg_license(algorithm)

Return the name of the software license, such as "MIT", applying to the package where the core algorithm for algorithm is implemented.

LearnAPI.doc_url — Function

LearnAPI.doc_url(algorithm)

Return a url where the core algorithm for algorithm is documented.

Returns "unknown" if the algorithm implementation has failed to overload the trait.

New implementations

Must return a string, such as "https://en.wikipedia.org/wiki/Decision_tree_learning".

LearnAPI.load_path — Function

LearnAPI.load_path(algorithm)

Return a string indicating where the struct for typeof(algorithm) can be found, beginning with the name of the package module defining it. For example, a return value of "FastTrees.LearnAPI.DecisionTreeClassifier" means the following julia code will return the algorithm type:

import FastTrees
FastTrees.LearnAPI.DecisionTreeClassifier

Returns "unknown" if the algorithm implementation has failed to overload the trait.

LearnAPI.is_wrapper — Function

LearnAPI.is_wrapper(algorithm)

Returns true if one or more properties (fields) of algorithm may themselves be algorithms, and false otherwise.

New implementations

This trait must be overloaded if one or more properties (fields) of algorithm may take algorithm values. Fallback return value is false.

The value of the trait must depend only on the type of algorithm.

LearnAPI.fit_keywords — Function

LearnAPI.fit_keywords(algorithm)

Return a list of keywords that can be provided to fit that correspond to metadata; metadata is for extra information pertaining to the data that is never iterated or subsampled. Examples, include target class weights and group lasso feature groupings. Further examples include feature names, and the pool of target classes, when these are not embedded in the data representation.

New implementations

If LearnAPI.fit(algorithm, ...) supports keyword arguments, then this trait must be overloaded, and otherwise not. Fallback returns ().

Here's a sample implementation for a classifier that implements a LearnAPI.fit method with signature fit(algorithm::MyClassifier, verbosity, X, y; class_weights=nothing):

LearnAPI.fit_keywords(::Any{<:MyClassifier}) = (:class_weights,)

or the shorthand

@trait MyClassifier fit_keywords=(:class_weights,)

LearnAPI.human_name — Function

LearnAPI.human_name(algorithm)

A human-readable string representation of typeof(algorithm). Primarily intended for auto-generation of documentation.

New implementations

Optional. A fallback takes the type name, inserts spaces and removes capitalization. For example, KNNRegressor becomes "knn regressor". Better would be to overload the trait to return "K-nearest neighbors regressor". Ideally, this is a "concrete" noun like "ridge regressor" rather than an "abstract" noun like "ridge regression".

LearnAPI.iteration_parameter — Function

LearnAPI.iteration_parameter(algorithm)

The name of the iteration parameter of algorithm, or nothing if the algorithm is not iterative.

New implementations

Implement if algorithm is iterative. Returns a symbol or nothing.

LearnAPI.fit_scitype — Function

LearnAPI.fit_scitype(algorithm)

Return an upper bound on the scitype of data guaranteeing it to work when training algorithm.

Specifically, if the return value is S and ScientificTypes.scitype(data) <: S, then the following low-level calls are allowed (assuming metadata is also valid and verbosity is an integer):

# apply data front-end:
data2, metadata2 = LearnAPI.reformat(algorithm, LearnAPI.fit, data...; metadata...)

# train:
LearnAPI.fit(algorithm, verbosity, data2...; metadata2...)

See also LearnAPI.fit_type, LearnAPI.fit_observation_scitype, LearnAPI.fit_observation_type.

New implementations

Optional. The fallback return value is Union{}. No more than one of the following should be overloaded for an algorithm type: LearnAPI.fit_scitype, LearnAPI.fit_type, LearnAPI.fit_observation_scitype, LearnAPI.fit_observation_type.

LearnAPI.fit_type — Function

LearnAPI.fit_type(algorithm)

Return an upper bound on the type of data guaranteeing it to work when training algorithm.

Specifically, if the return value is T and typeof(data) <: T, then the following low-level calls are allowed (assuming metadata is also valid and verbosity is an integer):

# apply data front-end:
data2, metadata2 = LearnAPI.reformat(algorithm, LearnAPI.fit, data...; metadata...)

# train:
LearnAPI.fit(algorithm, verbosity, data2...; metadata2...)

See also LearnAPI.fit_scitype, LearnAPI.fit_observation_type. LearnAPI.fit_observation_scitype

New implementations

Optional. The fallback return value is Union{}. No more than one of the following should be overloaded for an algorithm type: LearnAPI.fit_scitype, LearnAPI.fit_type, LearnAPI.fit_observation_scitype, LearnAPI.fit_observation_type.

LearnAPI.fit_observation_scitype — Function

LearnAPI.fit_observation_scitype(algorithm)

Return an upper bound on the scitype of observations guaranteed to work when training algorithm (independent of the type/scitype of the data container itself).

Specifically, denoting the type returned above by S, suppose a user supplies training data, data - typically a tuple, such as (X, y) - and valid metadata, metadata, and one computes

data2, metadata2 = LearnAPI.reformat(algorithm, LearnAPI.fit, data...; metadata...)

Then, assuming

ScientificTypes.scitype(LearnAPI.getobs(algorithm, LearnAPI.fit, data2, i)) <: S

for any valid index i, the following is guaranteed to work:

LearnAPI.fit(algorithm, verbosity, data2...; metadata2...)

See also See also LearnAPI.fit_type, LearnAPI.fit_scitype, LearnAPI.fit_observation_type.

New implementations

Optional. The fallback return value is Union{}. No more than one of the following should be overloaded for an algorithm type: LearnAPI.fit_scitype, LearnAPI.fit_type, LearnAPI.fit_observation_scitype, LearnAPI.fit_observation_type.

LearnAPI.fit_observation_type — Function

LearnAPI.fit_observation_type(algorithm)

Return an upper bound on the type of observations guaranteed to work when training algorithm (independent of the type/scitype of the data container itself).

Specifically, denoting the type returned above by T, suppose a user supplies training data, data - typically a tuple, such as (X, y) - and valid metadata, metadata, and one computes

data2, metadata2 = LearnAPI.reformat(algorithm, LearnAPI.fit, data...; metadata...)

Then, assuming

typeof(LearnAPI.getobs(algorithm, LearnAPI.fit, data2, i)) <: T

for any valid index i, the following is guaranteed to work:

LearnAPI.fit(algorithm, verbosity, data2...; metadata2...)

See also See also LearnAPI.fit_type, LearnAPI.fit_scitype, LearnAPI.fit_observation_scitype.

New implementations

Optional. The fallback return value is Union{}. No more than one of the following should be overloaded for an algorithm type: LearnAPI.fit_scitype, LearnAPI.fit_type, LearnAPI.fit_observation_scitype, LearnAPI.fit_observation_type.

LearnAPI.predict_input_scitype — Function

 LearnAPI.predict_input_scitype(algorithm)

Return an upper bound on the scitype of input data guaranteed to work with the predict operation.

Specifically, if S is the value returned and ScientificTypes.scitype(data) <: S, then the following low-level calls are allowed

 data2 = LearnAPI.reformat(algorithm, LearnAPI.predict, data...)
 LearnAPI.predict(algorithm, fitted_params, data2...)

Here fitted_params are the learned parameters returned by an appropriate call to LearnAPI.fit.

See also LearnAPI.predict_input_type.

New implementations

Implementation is optional. The fallback return value is Union{}. Should not be overloaded if LearnAPI.predict_input_type is overloaded.

LearnAPI.predict_output_scitype — Function

LearnAPI.predict_output_scitype(algorithm, kind_of_proxy::KindOfProxy)

Return an upper bound for the scitypes of predictions of the specified form where supported, and otherwise return Any. For example, if

ŷ, report = LearnAPI.predict(algorithm, LearnAPI.Distribution(), data...)

successfully returns (i.e., algorithm supports predictions of target probability distributions) then the following is guaranteed to hold:

scitype(ŷ) <: LearnAPI.predict_output_scitype(algorithm, LearnAPI.Distribution())

Note. This trait has a single-argument "convenience" version LearnAPI.predict_output_scitype(algorithm) derived from this one, which returns a dictionary keyed on target proxy types.

See also LearnAPI.KindOfProxy, LearnAPI.predict, LearnAPI.predict_input_scitype.

New implementations

Overloading the trait is optional. Here's a sample implementation for a supervised regressor type MyRgs that only predicts actual values of the target:

LearnAPI.predict(alogrithm::MyRgs, ::LearnAPI.LiteralTarget, data...) = ...
LearnAPI.predict_output_scitype(::MyRgs, ::LearnAPI.LiteralTarget) =
    AbstractVector{ScientificTypesBase.Continuous}

The fallback method returns Any.

LearnAPI.predict_output_scitype(algorithm)

Return a dictionary of upper bounds on the scitype of predictions, keyed on concrete subtypes of LearnAPI.KindOfProxy. Each of these subtypes respresents a different form of target prediction (LiteralTarget, Distribution, SurvivalFunction, etc) possibly supported by algorithm, but the existence of a key does not guarantee that form is supported.

As an example, if

ŷ, report = LearnAPI.predict(algorithm, LearnAPI.Distribution(), data...)

successfully returns (i.e., algorithm supports predictions of target probability distributions) then the following is guaranteed to hold:

scitype(ŷ) <: LearnAPI.predict_output_scitype(algorithm)[LearnAPI.Distribution]

See also LearnAPI.KindOfProxy, LearnAPI.predict, LearnAPI.predict_input_scitype.

New implementations

This single argument trait should not be overloaded. Instead, overload LearnAPI.predict_output_scitype(algorithm, kindofproxy). See above.

LearnAPI.predict_input_type — Function

LearnAPI.predict_input_type(algorithm)

Return an upper bound on the type of input data guaranteed to work with the predict operation.

Specifically, if T is the value returned and typeof(data) <: S, then the following low-level calls are allowed

data2 = LearnAPI.reformat(algorithm, LearnAPI.predict, data...)
LearnAPI.predict(algorithm, fitted_params, data2...)

Here fitted_params are the learned parameters returned by an appropriate call to LearnAPI.fit.

See also LearnAPI.predict_input_scitype.

New implementations

Implementation is optional. The fallback return value is Union{}. Should not be overloaded if LearnAPI.predict_input_scitype is overloaded.

LearnAPI.predict_output_type — Function

LearnAPI.predict_output_type(algorithm, kind_of_proxy::KindOfProxy)

Return an upper bound for the types of predictions of the specified form where supported, and otherwise return Any. For example, if

ŷ, report = LearnAPI.predict(algorithm, LearnAPI.Distribution(), data...)

successfully returns (i.e., algorithm supports predictions of target probability distributions) then the following is guaranteed to hold:

type(ŷ) <: LearnAPI.predict_output_type(algorithm, LearnAPI.Distribution())

Note. This trait has a single-argument "convenience" version LearnAPI.predict_output_type(algorithm) derived from this one, which returns a dictionary keyed on target proxy types.

See also LearnAPI.KindOfProxy, LearnAPI.predict, LearnAPI.predict_input_type.

New implementations

Overloading the trait is optional. Here's a sample implementation for a supervised regressor type MyRgs that only predicts actual values of the target:

LearnAPI.predict(alogrithm::MyRgs, ::LearnAPI.LiteralTarget, data...) = ...
LearnAPI.predict_output_type(::MyRgs, ::LearnAPI.LiteralTarget) =
    AbstractVector{ScientificTypesBase.Continuous}

The fallback method returns Any.

LearnAPI.predict_output_type(algorithm)

Return a dictionary of upper bounds on the type of predictions, keyed on concrete subtypes of LearnAPI.KindOfProxy. Each of these subtypes respresents a different form of target prediction (LiteralTarget, Distribution, SurvivalFunction, etc) possibly supported by algorithm, but the existence of a key does not guarantee that form is supported.

As an example, if

ŷ, report = LearnAPI.predict(algorithm, LearnAPI.Distribution(), data...)

successfully returns (i.e., algorithm supports predictions of target probability distributions) then the following is guaranteed to hold:

type(ŷ) <: LearnAPI.predict_output_type(algorithm)[LearnAPI.Distribution]

See also LearnAPI.KindOfProxy, LearnAPI.predict, LearnAPI.predict_input_type.

New implementations

This single argument trait should not be overloaded. Instead, overload LearnAPI.predict_output_type(algorithm, kindofproxy). See above.

LearnAPI.transform_input_scitype — Function

 LearnAPI.transform_input_scitype(algorithm)

Return an upper bound on the scitype of input data guaranteed to work with the transform operation.

Specifically, if S is the value returned and ScientificTypes.scitype(data) <: S, then the following low-level calls are allowed

 data2 = LearnAPI.reformat(algorithm, LearnAPI.transform, data...)
 LearnAPI.transform(algorithm, fitted_params, data2...)

Here fitted_params are the learned parameters returned by an appropriate call to LearnAPI.fit.

See also LearnAPI.transform_input_type.

New implementations

Implementation is optional. The fallback return value is Union{}. Should not be overloaded if LearnAPI.transform_input_type is overloaded.

LearnAPI.transform_output_scitype — Function

LearnAPI.transform_output_scitype(algorithm)

Return an upper bound on the scitype of the output of the transform operation.

Specifically, if S is the value returned, and if

output, report = LearnAPI.transform(algorithm, fitted_params, data...)

for suitable fitted_params and data, then

ScientificTypes.scitype(output) <: S

See also LearnAPI.transform_input_scitype.

New implementations

Implementation is optional. The fallback return value is Any.

LearnAPI.transform_input_type — Function

LearnAPI.transform_input_type(algorithm)

Return an upper bound on the type of input data guaranteed to work with the transform operation.

Specifically, if T is the value returned and typeof(data) <: S, then the following low-level calls are allowed

data2 = LearnAPI.reformat(algorithm, LearnAPI.transform, data...)
LearnAPI.transform(algorithm, fitted_params, data2...)

Here fitted_params are the learned parameters returned by an appropriate call to LearnAPI.fit.

See also LearnAPI.transform_input_scitype.

New implementations

Implementation is optional. The fallback return value is Union{}. Should not be overloaded if LearnAPI.transform_input_scitype is overloaded.

LearnAPI.transform_output_type — Function

LearnAPI.transform_output_type(algorithm)

Return an upper bound on the type of the output of the transform operation.

Specifically, if T is the value returned, and if

output, report = LearnAPI.transform(algorithm, fitted_params, data...)

for suitable fitted_params and data, then

typeof(output) <: T

See also LearnAPI.transform_input_type.

New implementations

Implementation is optional. The fallback return value is Any.