fit, update, update_observations, and update_features

Training

fit(algorithm, data; verbosity=1) -> model
fit(algorithm; verbosity=1) -> static_model 

A "static" algorithm is one that does not generalize to new observations (e.g., some clustering algorithms); there is no trainiing data and the algorithm is executed by predict or transform which receive the data. See example below.

When fit expects a tuple form of argument, data = (X1, ..., Xn), then the signature fit(algorithm, X1, ..., Xn) is also provided.

Updating

update(model, data; verbosity=1, param1=new_value1, param2=new_value2, ...) -> updated_model
update_observations(model, new_data; verbosity=1, param1=new_value1, ...) -> updated_model
update_features(model, new_data; verbosity=1, param1=new_value1, ...) -> updated_model

Data slurping forms are similarly provided for updating methods.

Typical workflows

Supposing Algorithm is some supervised classifier type, with an iteration parameter n:

algorithm = Algorithm(n=100)
model = fit(algorithm, (X, y)) # or `fit(algorithm, X, y)`

# Predict probability distributions:
ŷ = predict(model, Distribution(), Xnew) 

# Inspect some byproducts of training:
LearnAPI.feature_importances(model)

# Add 50 iterations and predict again:
model = update(model; n=150)
predict(model, Distribution(), X)

A static algorithm (no "learning")

# Apply some clustering algorithm which cannot be generalized to new data:
model = fit(algorithm) # no training data
labels = predict(model, LabelAmbiguous(), X) # may mutate `model`

# Or, in one line:
labels = predict(algorithm, LabelAmbiguous(), X)

# But two-line version exposes byproducts of the clustering algorithm (e.g., outliers):
LearnAPI.extras(model)

Implementation guide

Training

method	fallback	compulsory?

fit(algorithm, data; verbosity=1)	ignores data and applies signature below	yes, unless static

fit(algorithm; verbosity=1)	none	no, unless static

Updating

method	fallback	compulsory?

update(model, data; verbosity=1, hyperparameter_updates...)	none	no

update_observations(model, data; verbosity=1, hyperparameter_updates...)	none	no

update_features(model, data; verbosity=1, hyperparameter_updates...)	none	no

There are some contracts regarding the behaviour of the update methods, as they relate to a previous fit call. Consult the document strings for details.

Reference

LearnAPI.fit — Function

fit(algorithm, data; verbosity=1)
fit(algorithm; verbosity=1)

Execute the algorithm with configuration algorithm using the provided training data, returning an object, model, on which other methods, such as predict or transform, can be dispatched. LearnAPI.functions(algorithm) returns a list of methods that can be applied to either algorithm or model.

The second signature is provided by algorithms that do not generalize to new observations ("static" algorithms). In that case, transform(model, data) or predict(model, ..., data) carries out the actual algorithm execution, writing any byproducts of that operation to the mutable object model returned by fit.

Whenever fit expects a tuple form of argument, data = (X1, ..., Xn), then the signature fit(algorithm, X1, ..., Xn) is also provided.

For example, a supervised classifier will typically admit this workflow:

model = fit(algorithm, (X, y)) # or `fit(algorithm, X, y)`
ŷ = predict(model, Xnew)

Use verbosity=0 for warnings only, and -1 for silent training.

See also predict, transform, inverse_transform, LearnAPI.functions, obs.

Extended help

New implementations

Implementation is compulsory. The signature must include verbosity. A fallback for the first signature calls the second, ignoring data:

fit(algorithm, data; kwargs...) = fit(algorithm; kwargs...)

Fallbacks also provide the data slurping versions.

Assumptions about data

By default, it is assumed that data supports the LearnAPI.RandomAccess interface; this includes all matrices, with observations-as-columns, most tables, and tuples thereof). See LearnAPI.RandomAccess for details. If this is not the case then an implementation must either: (i) overload obs to articulate how provided data can be transformed into a form that does support LearnAPI.RandomAccess; or (ii) overload the trait LearnAPI.data_interface to specify a more relaxed data API. Refer to document strings for details.

LearnAPI.update — Function

update(model, data; verbosity=1, hyperparam_replacements...)

Return an updated version of the model object returned by a previous fit or update call, but with the specified hyperparameter replacements, in the form p1=value1, p2=value2, ....

Provided that data is identical with the data presented in a preceding fit call, as in the example below, execution is semantically equivalent to the call fit(algorithm, data), where algorithm is LearnAPI.algorithm(model) with the specified replacements. In some cases (typically, when changing an iteration parameter) there may be a performance benefit to using update instead of retraining ab initio.

If data differs from that in the preceding fit or update call, then behaviour is algorithm-specific.

algorithm = MyForest(ntrees=100)

# train with 100 trees:
model = fit(algorithm, data)

# add 50 more trees:
model = update(model, data; ntrees=150)

See also fit, update_observations, update_features.

New implementations

Implementation is optional. The signature must include verbosity. If implemented, you must include :(LearnAPI.update) in the tuple returned by the LearnAPI.functions trait.

LearnAPI.update_observations — Function

update_observations(model, new_data; verbosity=1, parameter_replacements...)

Return an updated version of the model object returned by a previous fit or update call given the new observations present in new_data. One may additionally specify hyperparameter replacements in the form p1=value1, p2=value2, ....

When following the call fit(algorithm, data), the update call is semantically equivalent to retraining ab initio using a concatenation of data and new_data, provided there are no hyperparameter replacements. Behaviour is otherwise algorithm-specific.

algorithm = MyNeuralNetwork(epochs=10, learning_rate=0.01)

# train for ten epochs:
model = fit(algorithm, data)

# train for two more epochs using new data and new learning rate:
model = update_observations(model, new_data; epochs=2, learning_rate=0.1)

See also fit, update, update_features.

Extended help

New implementations

Implementation is optional. The signature must include verbosity. If implemented, you must include :(LearnAPI.update_observations) in the tuple returned by the LearnAPI.functions trait.

LearnAPI.update_features — Function

update_features(model, new_data; verbosity=1, parameter_replacements...)

Return an updated version of the model object returned by a previous fit or update call given the new features encapsulated in new_data. One may additionally specify hyperparameter replacements in the form p1=value1, p2=value2, ....

When following the call fit(algorithm, data), the update call is semantically equivalent to retraining ab initio using a concatenation of data and new_data, provided there are no hyperparameter replacements. Behaviour is otherwise algorithm-specific.

See also fit, update, update_features.

Extended help

New implementations

Implementation is optional. The signature must include verbosity. If implemented, you must include :(LearnAPI.update_features) in the tuple returned by the LearnAPI.functions trait.