Fit, update! and ingest!

Summary. Algorithms that learn, i.e., generalize to new data, must overload fit; the fallback performs no operation and returns all nothing. Implement update! if certain hyperparameter changes do not necessitate retraining from scratch (e.g., increasing an iteration parameter). Implement ingest! to implement incremental learning. All training methods implemented must be named in the return value of the functions trait.

method	fallback	compulsory?	requires

LearnAPI.fit	does nothing, returns (nothing, nothing, nothing)	no

LearnAPI.update!	calls fit	no	LearnAPI.fit

LearnAPI.ingest!	none	no	LearnAPI.fit

All three methods above return a triple (fitted_params, state, report) whose components are explained under LearnAPI.fit below. Items that might be returned in report include: feature rankings/importances, SVM support vectors, clustering centers, methods for visualizing training outcomes, methods for saving learned parameters in a custom format, degrees of freedom, deviances. Precisely what report includes might be controlled by hyperparameters (algorithm properties) especially if there is a performance cost to it's inclusion.

Implement fit unless all operations, such as predict and transform, ignore their fitted_params argument (which will be nothing). This is the case for many algorithms that have hyperparameters, but do not generalize to new data, such as a basic DBSCAN clustering algorithm.

The update! method is intended for all subsequent calls to train an algorithm using the same observations, but with possibly altered hyperparameters (algorithm argument). A fallback implementation simply calls fit. The main use cases for implementing update are:

warm-restarting iterative algorithms

"smart" training of composite algorithms, such as linear pipelines; here "smart" means that hyperparameter changes only trigger the retraining of downstream components.

The ingest! method supports incremental learning (same hyperparameters, but new training observations). Like update!, it depends on the output a preceding fit or ingest! call.

LearnAPI.fit — Function

LearnAPI.fit(algorithm, verbosity, data...; metadata...)

Perform training associated with algorithm using the provided data and metadata. With the exception of warnings, training will be silent if verbosity == 0. Lower values should suppress warnings; any integer ought to be admissible. Here:

algorithm is a property-accessible object whose properties are the hyperparameters of some ML/statistical algorithm.

data is a tuple of data objects with a common number of observations, for example, data = (X, y, w) where X is a table of features, y is a target vector with the same number of rows, and w a vector of per-observation weights.

metadata is for extra information pertaining to the data that is never iterated or subsampled. Examples, include target class weights and group lasso feature groupings. Further examples include feature names, and the pool of target classes, when these are not embedded in the data representation. To see the keyword names for metadata supported by algorithm, do LearnAPI.fit_keywords(algorithm). "

Return value

Returns a tuple (fitted_params, state, report) where:

The fitted_params is the algorithm's learned parameters (eg, the coefficients in a linear algorithm) in a form understood by operations. An operation is a method dispatched on an algorithm, associated learned parameters, and data. The LearnAPI operations are: LearnAPI.predict, LearnAPI.transform, LearnAPI.inverse_transform. If some training outcome of user-interest is not needed for operations, it should be part of report instead (see below).

The state is for passing to LearnAPI.update! or LearnAPI.ingest!. For algorithms that implement neither, state should be nothing.

The report records byproducts of training not in the fitted_params, such as feature rankings, or out-of-sample estimates of performance.

New implementations

Overloading this method for new algorithms is optional. A fallback performs no computation, returning (nothing, nothing, nothing).

See the LearnAPI.jl documentation for the detailed requirements of LearnAPI.jl algorithm objects.

In LearnAPI, data is any tuple of objects sharing a common number of "observations. "

Note

The method is not permitted to mutate algorithm. In particular, if algorithm has a random number generator as a hyperparameter (property) then it must be copied before use.

If implemented, you must include :fit in the tuple returned by the LearnAPI.functions trait.

If supporting metadata, you must also implement LearnAPI.fit_keywords to list the supported keyword argument names (e.g., class_weights).

See also LearnAPI.update!, LearnAPI.ingest!.

LearnAPI.update! — Function

LearnAPI.update!(algorithm, verbosity, fitted_params, state, data...; metadata...)

Based on the values of state, and fitted_params returned by a preceding call to LearnAPI.fit, LearnAPI.ingest!, or LearnAPI.update!, update a algorithm's fitted parameters, returning new (or mutated) state and fitted_params.

Intended for retraining when the training data has not changed, but algorithm properties (hyperparameters) may have changed, e.g., when increasing an iteration parameter. Specifically, the assumption is that data and metadata have the same values seen in the most recent call to fit/update!/ingest!.

For incremental training (same algorithm, new data) see instead LearnAPI.ingest!.

Return value

Same as LearnAPI.fit, namely a tuple (fitted_params, state, report). See LearnAPI.fit for details.

New implementations

Overloading this method is optional. A fallback calls LearnAPI.fit:

LearnAPI.update!(algorithm, verbosity, fitted_params, state, data...; metadata...) =
    fit(algorithm, verbosity, data; metadata...)

If implemented, you must include :fit in the tuple returned by the LearnAPI.functions trait.

In LearnAPI, data is any tuple of objects sharing a common number of "observations. "

The most common use case is continuing training of an iterative algorithm: state is simply a copy of the algorithm used in the last training call (fit, update! or ingest!) and this will include the current number of iterations as a property. If algorithm and state differ only in the number of iterations (e.g., epochs in a neural network), which has increased, then the fitted parameters (weights) are updated, rather than computed from scratch. Otherwise, update! simply calls fit, to force retraining from scratch.

It is permitted to return mutated versions of state and fitted_params.

Note

The method is not permitted to mutate algorithm. In particular, if algorithm has a random number generator as a hyperparameter (property) then it must be copied before use.

See also LearnAPI.fit, LearnAPI.ingest!.

LearnAPI.ingest! — Function

LernAPI.ingest!(algorithm, verbosity, fitted_params, state, data...; metadata...)

For an algorithm that supports incremental learning, update the fitted parameters using data, which has typically not been seen before. The arguments state and fitted_params are the output of a preceding call to LearnAPI.fit, LearnAPI.ingest!, or LearnAPI.update!, of which mutated or new versions are returned.

For updating fitted parameters using the same data but new hyperparameters, see instead LearnAPI.update!.

For training an algorithm with new hyperparameters but unchanged data, see instead LearnAPI.update!.

Return value

Same as LearnAPI.fit, namely a tuple (fitted_params, state, report). See LearnAPI.fit for details.

New implementations

Implementing this method is optional. It has no fallback.

If implemented, you must include :fit in the tuple returned by the LearnAPI.functions trait.

Note

The method is not permitted to mutate algorithm. In particular, if algorithm has a random number generator as a hyperparameter (property) then it must be copied before use.

See also LearnAPI.fit, LearnAPI.update!.