Kinds of learner

LearnAPI.KindOfLearner — Type

LearnAPI.KindOfLearner

Abstract type whose instances are the possible values of LearnAPI.kind_of(learner). All instances of this type, and brief indications of their interpretation, appear below.

LearnAPI.Descriminative(): A typical workflow looks like:

model = fit(learner, data)
predict(learner, new_data)
# or
transform(learner, new_data)

LearnAPI.Static(): A typical workflow looks like:

model = fit(learner)
predict(model, data) # may mutate `model` to record byproducts of computation
# or
transform(model, data)

LearnAPI.Generative(): A typical workflow looks like:

model = fit(learner, data)
predict(learner) # e.g., returns a single probability distribution

For precise details, refer to the document strings for LearnAPI.Descriminative, LearnAPI.Static, and LearnAPI.Generative.

LearnAPI.Descriminative — Type

LearnAPI.Descriminative

Type with a single instance, LearnAPI.Descriminative().

If LearnAPI.kind_of(learner)== LearnAPI.Descriminative(), then the only possible signatures of fit, predict and transform are those appearing below, or variations on these in which keyword arguments are also supplied:

model = fit(learner, data)
predict(model, new_data)
predict(model, kop::KindOfProxy, new_data)
transform(model, new_data)

and the one-line convenience forms

predict(learner, data)
predict(learner, kop::KindOfProxy, new_data)
transform(learner, data)

See also LearnAPI.Static, LearnAPI.Generative.

LearnAPI.Static — Type

LearnAPI.Static

Type with a single instance, LearnAPI.Static().

If LearnAPI.kind_of(learner)== LearnAPI.Static(), then the only possible signatures of fit, predict and transform are those appearing below, or variations on these in which keyword arguments are also supplied:

model = fit(learner)   # no `data` argument
predict(model, data)                   # may mutate `model`
predict(model, kop::KindOfProxy, data) # may mutate `model`
transform(model, data)                 # may mutate `model`

and the one-line convenience forms

predict(learner, data)
predict(learner, kop::KindOfProxy)
transform(learner, data)

See also LearnAPI.Descriminative, LearnAPI.Generative.

LearnAPI.Generative — Type

LearnAPI.Generative

Type with a single instance, LearnAPI.Generative().

If LearnAPI.kind_of(learner)== LearnAPI.Generative(), then the only possible signatures of fit, predict and transform are those appearing below, or variations on these in which keyword arguments are also supplied:

model = fit(learner, data)
predict(model)                    # no `newdata` argument
predict(model, kop::KindOfProxy)  # no `newdata` argument
transform(model)                  # no `newdata` argument

and the one-line convenience forms

predict(learner, data)
predict(learner, kop::KindOfProxy, data)
transform(learner, data)

See also LearnAPI.Descriminative, LearnAPI.Static.