Model Traits

Summary. Traits allow one to promise particular behaviour for a model, such as: This model supports per-observation weights, which must appear as the third argument of fit, or This model predicts probability distributions for the target, or This model's transform method predicts Real vectors.

Traits are optional, except where required by the implementation of some LearnAPI method and documented in that method's docstring. Unless a model has no functionality whatsoever, LearnAPI.functions(model) will need to be overloaded.

Traits are often called on instances but are frequently defined on model types, as in

LearnAPI.is_pure_julia(::Type{<:MyModelType}) = true

which has the shorthand

@trait MyModelType is_pure_julia=true

So, for convenience, every trait t is provided the fallback implementation

t(model) = t(typeof(model))

This means LearnAPI.is_pure_julia(model) = true whenever model isa MyModelType in the above example.

Every trait has a global fallback implementation for ::Type.

Traits that vary from instance to instance of the same type are discouraged, except in the case of composite models (is_wrapper(model) = true) where this is unavoidable. One reason for this is so one can associate with each model type a unique set of trait-based "model metadata" for inclusion in searchable model databases. This requirement occasionally requires that an existing model implementation be split into several separate LearnAPI implementations (e.g., one for regression and another for classification).

Ordinary traits are available for overloading by an new model implementation. Derived traits are not.

Ordinary traits

In the examples column of the table below, Table, Continuous, Sampleable are names owned by the package ScientificTypesBase.jl.

trait	fallback value	return value	example

LearnAPI.functions(model)	()	implemented LearnAPI functions (traits excluded)	(:fit, :predict)

LearnAPI.predict_proxy(model)	LearnAPI.None()	form of target proxy output by predict	LearnAPI.Distribution()

LearnAPI.predict_joint_proxy(model)	LearnAPI.None()	form of target proxy output by predict_joint	LearnAPI.Distribution()

LearnAPI.position_of_target(model)	0	† the positional index of the target in data in fit(..., data...; metadata) calls	2

LearnAPI.position_of_weights(model)	0	† the positional index of per-observation weights in data in fit(..., data...; metadata)	3

LearnAPI.descriptors(model)	()	lists one or more suggestive model descriptors from LearnAPI.descriptors()	(:classifier, :probabilistic)

LearnAPI.is_pure_julia(model)	false	is true if implementation is 100% Julia code	true

LearnAPI.pkg_name(model)	"unknown"	name of package providing core algorithm (may be different from package providing LearnAPI.jl implementation)	"DecisionTree"

LearnAPI.pkg_license(model)	"unknown"	name of license of package providing core algorithm	"MIT"

LearnAPI.doc_url(model)	"unknown"	url providing documentation of the core algorithm	"https://en.wikipedia.org/wiki/Decision_tree_learning"

LearnAPI.load_path(model)	"unknown"	a string indicating where the struct for typeof(model) is defined, beginning with name of package providing implementation	FastTrees.LearnAPI.DecisionTreeClassifier

LearnAPI.is_wrapper(model)	false	is true if one or more properties (fields) of model may be a model	true

LearnAPI.human_name(model)	type name with spaces	human name for the model; should be a noun	"elastic net regressor"

LearnAPI.iteration_parameter(model)	nothing	symbolic name of an iteration parameter	:epochs

LearnAPI.fit_keywords(model)	()	tuple of symbols for keyword arguments accepted by fit (corresponding to metadata)	(:class_weights,)

LearnAPI.fit_scitype(model)	Union{}	upper bound on scitype(data) in fit(model, verbosity, data...)††	Tuple{Table(Continuous), AbstractVector{Continuous}}

LearnAPI.fit_observation_scitype(model)	Union{}	upper bound on scitype(observation) for observation in data and data in fit(model, verbosity, data...)††	Tuple{AbstractVector{Continuous}, Continuous}

LearnAPI.fit_type(model)	Union{}	upper bound on type(data) in fit(model, verbosity, data...)††	Tuple{AbstractMatrix{<:Real}, AbstractVector{<:Real}}

LearnAPI.fit_observation_type(model)	Union{}	upper bound on type(observation) for observation in data and data in fit(model, verbosity, data...)*	Tuple{AbstractVector{<:Real}, Real}

LearnAPI.predict_input_scitype(model)	Union{}	upper bound on scitype(data) in predict(model, fitted_params, data...)††	Table(Continuous)

LearnAPI.predict_output_scitype(model)	Any	upper bound on scitype(first(predict(model, ...)))	AbstractVector{Continuous}

LearnAPI.predict_input_type(model)	Union{}	upper bound on typeof(data) in predict(model, fitted_params, data...)††	AbstractMatrix{<:Real}

LearnAPI.predict_output_type(model)	Any	upper bound on typeof(first(predict(model, ...)))	AbstractVector{<:Real}

LearnAPI.predict_joint_input_scitype(model)	Union{}	upper bound on scitype(data) in predict_joint(model, fitted_params, data...)††	Table(Continuous)

LearnAPI.predict_joint_output_scitype(model)	Any	upper bound on scitype(first(predict_joint(model, ...)))	Sampleable{<:AbstractVector{Continuous}}

LearnAPI.predict_joint_input_type(model)	Union{}	upper bound on typeof(data) in predict_joint(model, fitted_params, data...)††	AbstractMatrix{<:Real}

LearnAPI.predict_joint_output_type(model)	Any	upper bound on typeof(first(predict_joint(model, ...)))	Distributions.Sampleable{Distributions.Multivariate,Distributions.Continuous}

LearnAPI.transform_input_scitype(model)	Union{}	upper bound on scitype(data) in transform(model, fitted_params, data...)††	Table(Continuous)

LearnAPI.transform_output_scitype(model)	Any	upper bound on scitype(first(transform(model, ...)))	Table(Continuous)

LearnAPI.transform_input_type(model)	Union{}	upper bound on typeof(data) in transform(model, fitted_params, data...)††	AbstractMatrix{<:Real}}

LearnAPI.transform_output_type(model)	Any	upper bound on typeof(first(transform(model, ...)))	AbstractMatrix{<:Real}

LearnAPI.inverse_transform_input_scitype(model)	Union{}	upper bound on scitype(data) in inverse_transform(model, fitted_params, data...)††	Table(Continuous)

LearnAPI.inverse_transform_output_scitype(model)	Any	upper bound on scitype(first(inverse_transform(model, ...)))	Table(Continuous)

LearnAPI.inverse_transform_input_type(model)	Union{}	upper bound on typeof(data) in inverse_transform(model, fitted_params, data...)††	AbstractMatrix{<:Real}

LearnAPI.inverse_transform_output_type(model)	Any	upper bound on typeof(first(inverse_transform(model, ...)))	AbstractMatrix{<:Real}

† If the value is 0, then the variable in boldface type is not supported and not expected to appear in data. If length(data) is less than the trait value, then data is understood to exclude the variable, but note that fit can have multiple signatures of varying lengths, as in fit(model, verbosity, X, y) and fit(model, verbosity, X, y, w). A non-zero value is a promise that fit includes a signature of sufficient length to include the variable.

†† Assuming no optional data interface is implemented. See docstring for the general case.

Derived Traits

The following convenience methods are provided but intended for overloading:

trait	return value	example

LearnAPI.name(model)	model type name as string	"PCA"

LearnAPI.is_model(model)	true if functions(model) is not empty	true

Reference

LearnAPI.functions — Function

LearnAPI.functions(model)

Return a tuple of symbols, such as (:fit, :predict), corresponding to LearnAPI methods specifically implemented for objects having the same type as model. If non-empty, this also guarantees model is a model, in the LearnAPI sense. See the Reference section of the manual for details.

New model implementations

Every LearnAPI method that is not a trait and which is specifically implemented for typeof(model) must be included in the return value of this trait. Specifically, the return value is a tuple of symbols from this list: :fit, :update!, :ingest!, :predict, :predict_joint, :transform, :inverse_transform, :features_importances, :training_labels, :training_losses, :training_scores. To regenerate this list, do LearnAPI.functions().

See also LearnAPI.Model.

LearnAPI.predict_proxy — Function

LearnAPI.predict_proxy(model)

Returns an object with abstract type LearnAPI.TargetProxy indicating the kind of proxy for the target returned by the predict method, when called on model and some data. For example, a value of LearnAPI.Distribution() means that predict returns probability distributions, rather than actual values of the target. (LearnAPI.predict also returns a report as second value). A value of LearnAPI.TrueTarget() indicates that ordinary (non-proxy) target values are returned. A value of LearnAPI.None(), implies the output of predict has no declared relationship with any target variable.

New implementations

A model with a concept of "target" must overload this trait. The fallback return value is LearnAPI.None().

For more on target variables and target proxies, refer to the "Predict and Other Operations" section of the LearnAPI documentation.

The trait must return a lone instance T() for some subtype T <: LearnAPI.TargetProxy. Here's a sample implementation for a supervised model where predictions are ordinary values of the target variable:

@trait MyNewModel predict_proxy = LearnAPI.TrueTarget()

which is shorthand for

LearnAPI.predict_proxy(::Type{<:MyNewModelType}) = LearnAPI.TrueTarget()

LearnAPI.predict_joint_proxy — Function

LearnAPI.predict_joint_proxy(model)

Returns an object with abstract type LearnAPI.TargetProxy indicating the kind of proxy for the target returned by the predict_joint method, when called on model and some data. For example, a value of LearnAPI.Distribution() means that predict_joint returns a probability distribution, rather than, say a merely sampleable object.

New implementations

Any model implementing LearnAPI.predict_joint must overload this trait.

For more on target variables and target proxies, refer to the LearnAPI documentation.

The possible return values for this trait are: LearnAPI.JointSampleable(), LearnAPI.JointDistribution() and LearnAPI.JointLogDistribution().

Here's a sample implementation:

@trait MyNewModel predict_joint_proxy = LearnAPI.JointDistribution()

which is shorthand for

LearnAPI.predict_joint_proxy(::Type{<:MyNewModelType}) = LearnAPI.JointDistribution()

LearnAPI.position_of_target — Function

LearnAPI.position_of_target(model)

Return the expected position of the target variable within data in calls of the form LearnAPI.fit(model, verbosity, data...).

If this number is 0, then no target is expected. If this number exceeds length(data), then data is understood to exclude the target variable.

LearnAPI.position_of_weights — Function

LearnAPI.position_of_weights(model)

Return the expected position of per-observation weights within data in calls of the form LearnAPI.fit(model, verbosity, data...).

If this number is 0, then no weights are expected. If this number exceeds length(data), then data is understood to exclude weights, which are assumed to be uniform.

LearnAPI.descriptors — Function

LearnAPI.descriptors(model)

Lists one or more suggestive model descriptors from this list: :regression, :classification, :clustering, :gradient_descent, :iterative_model, :incremental_model, :dimension_reduction, :transformer, :static_transformer, :missing_value_imputer, :ensemble_model, :wrapper, :time_series_forecaster, :time_series_classifier, :survival_analysis, :distribution_fitter, :Bayesian_model, :outlier_detection, :collaborative_filtering, :text_analysis, :audio_analysis, :natural_language_processing, :image_processing (do LearnAPI.descriptors() to reproduce).

Warning

The value of this trait guarantees no particular behavior. The trait is intended for informal classification purposes only.

New model implementations

This trait should return a tuple of symbols, as in (:classifier, :probabilistic).

LearnAPI.is_pure_julia — Function

LearnAPI.is_pure_julia(model)

Returns true if training model requires evaluation of pure Julia code only.

New model implementations

The fallback is false.

LearnAPI.pkg_name — Function

LearnAPI.pkg_name(model)

Return the name of the package module which supplies the core training algorithm for model. This is not necessarily the package providing the LearnAPI interface.

Returns "unknown" if the model implementation has failed to overload the trait.

New model implementations

Must return a string, as in "DecisionTree".

LearnAPI.pkg_license — Function

LearnAPI.pkg_license(model)

Return the name of the software license, such as "MIT", applying to the package where the core algorithm for model is implemented.

LearnAPI.doc_url — Function

LearnAPI.doc_url(model)

Return a url where the core algorithm for model is documented.

Returns "unknown" if the model implementation has failed to overload the trait.

New model implementations

Must return a string, such as "https://en.wikipedia.org/wiki/Decision_tree_learning".

LearnAPI.load_path — Function

LearnAPI.load_path(model)

Return a string indicating where the struct for typeof(model) can be found, beginning with the name of the package module defining it. For example, a return value of "FastTrees.LearnAPI.DecisionTreeClassifier" means the following julia code will return the model type:

import FastTrees
FastTrees.LearnAPI.DecisionTreeClassifier

Returns "unknown" if the model implementation has failed to overload the trait.

LearnAPI.is_wrapper — Function

LearnAPI.is_wrapper(model)

Returns true if one or more properties (fields) of model may themselves be models, and false otherwise.

New model implementations

This trait must be overloaded if one or more properties (fields) of model may take model values. Fallback return value is false.

The value of the trait must depend only on the type of model.

LearnAPI.fit_keywords — Function

LearnAPI.fit_keywords(model)

Return a list of keywords that can be provided to fit that correspond to metadata; metadata is for extra information pertaining to the data that is never iterated or subsampled, eg., weights for target classes, or feature names (if these are not embedded in the representation of data). Another example would be feature groupings in the group lasso algorithm.

New model implementations

If LearnAPI.fit(model, ...) supports keyword arguments, then this trait must be overloaded, and otherwise not. Fallback returns ().

Here's a sample implementation for a classifier that implements a LearnAPI.fit method with signature fit(model::MyClassifier, verbosity, X, y; class_weights=nothing):

LearnAPI.fit_keywords(::Type{<:MyClassifier}) = (:class_weights,)

or the shorthand

@trait MyClassifier fit_keywords=(:class_weights,)

LearnAPI.human_name — Function

LearnAPI.human_name(model)

A human-readable string representation of typeof(model). Primarily intended for auto-generation of documentation.

New model implementations

Optional. A fallback takes the type name, inserts spaces and removes capitalization. For example, KNNRegressor becomes "knn regressor". Better would be to overload the trait to return "K-nearest neighbors regressor". Ideally, this is a "concrete" noun like "ridge regressor" rather than an "abstract" noun like "ridge regression".

LearnAPI.iteration_parameter — Function

LearnAPI.iteration_parameter(model)

The name of the iteration parameter of model, or nothing if the model is not iterative.

New model implementations

Implement if model is iterative. Returns a symbol or nothing.

LearnAPI.fit_scitype — Function

LearnAPI.fit_scitype(model)

Return an upper bound on the scitype of data guaranteeing it to work when training model.

Specifically, if the return value is S and ScientificTypes.scitype(data) <: S, then the following low-level calls are allowed (assuming metadata is also valid and verbosity is an integer):

# apply data front-end:
data2, metadata2 = LearnAPI.reformat(model, LearnAPI.fit, data...; metadata...)

# train:
LearnAPI.fit(model, verbosity, data2...; metadata2...)

See also LearnAPI.fit_type, LearnAPI.fit_observation_scitype, LearnAPI.fit_observation_type.

New model implementations

Optional. The fallback return value is Union{}. No more than one of the following should be overloaded for a model type: LearnAPI.fit_scitype, LearnAPI.fit_type, LearnAPI.fit_observation_scitype, LearnAPI.fit_observation_type.

LearnAPI.fit_type — Function

LearnAPI.fit_type(model)

Return an upper bound on the type of data guaranteeing it to work when training model.

Specifically, if the return value is T and typeof(data) <: T, then the following low-level calls are allowed (assuming metadata is also valid and verbosity is an integer):

# apply data front-end:
data2, metadata2 = LearnAPI.reformat(model, LearnAPI.fit, data...; metadata...)

# train:
LearnAPI.fit(model, verbosity, data2...; metadata2...)

See also LearnAPI.fit_scitype, LearnAPI.fit_observation_type. LearnAPI.fit_observation_scitype

New model implementations

Optional. The fallback return value is Union{}. No more than one of the following should be overloaded for a model type: LearnAPI.fit_scitype, LearnAPI.fit_type, LearnAPI.fit_observation_scitype, LearnAPI.fit_observation_type.

LearnAPI.fit_observation_scitype — Function

LearnAPI.fit_observation_scitype(model)

Return an upper bound on the scitype of observations guaranteed to work when training model (independent of the type/scitype of the data container itself).

Specifically, denoting the type returned above by S, suppose a user supplies training data, data - typically a tuple, such as (X, y) - and valid metadata, metadata, and one computes

data2, metadata2 = LearnAPI.reformat(model, LearnAPI.fit, data...; metadata...)

Then, assuming

ScientificTypes.scitype(LearnAPI.getobs(model, LearnAPI.fit, data2, i)) <: S

for any valid index i, the following is guaranteed to work:

LearnAPI.fit(model, verbosity, data2...; metadata2...)

See also See also LearnAPI.fit_type, LearnAPI.fit_scitype, LearnAPI.fit_observation_type.

New model implementations

Optional. The fallback return value is Union{}. No more than one of the following should be overloaded for a model type: LearnAPI.fit_scitype, LearnAPI.fit_type, LearnAPI.fit_observation_scitype, LearnAPI.fit_observation_type.

LearnAPI.fit_observation_type — Function

LearnAPI.fit_observation_type(model)

Return an upper bound on the type of observations guaranteed to work when training model (independent of the type/scitype of the data container itself).

Specifically, denoting the type returned above by T, suppose a user supplies training data, data - typically a tuple, such as (X, y) - and valid metadata, metadata, and one computes

data2, metadata2 = LearnAPI.reformat(model, LearnAPI.fit, data...; metadata...)

Then, assuming

typeof(LearnAPI.getobs(model, LearnAPI.fit, data2, i)) <: T

for any valid index i, the following is guaranteed to work:

LearnAPI.fit(model, verbosity, data2...; metadata2...)

See also See also LearnAPI.fit_type, LearnAPI.fit_scitype, LearnAPI.fit_observation_scitype.

New model implementations

Optional. The fallback return value is Union{}. No more than one of the following should be overloaded for a model type: LearnAPI.fit_scitype, LearnAPI.fit_type, LearnAPI.fit_observation_scitype, LearnAPI.fit_observation_type.

LearnAPI.predict_input_scitype — Function

 LearnAPI.predict_input_scitype(model)

Return an upper bound on the scitype of input data guaranteed to work with the predict operation.

Specifically, if S is the value returned and ScientificTypes.scitype(data) <: S, then the following low-level calls are allowed

 data2 = LearnAPI.reformat(model, LearnAPI.predict, data...)
 LearnAPI.predict(model, fitted_params, data2...)

Here fitted_params are the learned parameters returned by an appropriate call to LearnAPI.fit.

See also LearnAPI.predict_input_type.

New model implementations

Implementation is optional. The fallback return value is Union{}. Should not be overloaded if LearnAPI.predict_input_type is overloaded.

LearnAPI.predict_output_scitype — Function

LearnAPI.predict_output_scitype(model)

Return an upper bound on the scitype of the output of the predict operation.

Specifically, if S is the value returned, and if

output, report = LearnAPI.predict(model, fitted_params, data...)

for suitable fitted_params and data, then

ScientificTypes.scitype(output) <: S

See also LearnAPI.predict_input_scitype.

New model implementations

Implementation is optional. The fallback return value is Any.

LearnAPI.predict_input_type — Function

LearnAPI.predict_input_type(model)

Return an upper bound on the type of input data guaranteed to work with the predict operation.

Specifically, if T is the value returned and typeof(data) <: S, then the following low-level calls are allowed

data2 = LearnAPI.reformat(model, LearnAPI.predict, data...)
LearnAPI.predict(model, fitted_params, data2...)

Here fitted_params are the learned parameters returned by an appropriate call to LearnAPI.fit.

See also LearnAPI.predict_input_scitype.

New model implementations

Implementation is optional. The fallback return value is Union{}. Should not be overloaded if LearnAPI.predict_input_scitype is overloaded.

LearnAPI.predict_output_type — Function

LearnAPI.predict_output_type(model)

Return an upper bound on the type of the output of the predict operation.

Specifically, if T is the value returned, and if

output, report = LearnAPI.predict(model, fitted_params, data...)

for suitable fitted_params and data, then

typeof(output) <: T

See also LearnAPI.predict_input_type.

New model implementations

Implementation is optional. The fallback return value is Any.

LearnAPI.predict_joint_input_scitype — Function

 LearnAPI.predict_joint_input_scitype(model)

Return an upper bound on the scitype of input data guaranteed to work with the predict_joint operation.

Specifically, if S is the value returned and ScientificTypes.scitype(data) <: S, then the following low-level calls are allowed

 data2 = LearnAPI.reformat(model, LearnAPI.predict_joint, data...)
 LearnAPI.predict_joint(model, fitted_params, data2...)

Here fitted_params are the learned parameters returned by an appropriate call to LearnAPI.fit.

See also LearnAPI.predict_joint_input_type.

New model implementations

Implementation is optional. The fallback return value is Union{}. Should not be overloaded if LearnAPI.predict_joint_input_type is overloaded.

LearnAPI.predict_joint_output_scitype — Function

LearnAPI.predict_joint_output_scitype(model)

Return an upper bound on the scitype of the output of the predict_joint operation.

Specifically, if S is the value returned, and if

output, report = LearnAPI.predict_joint(model, fitted_params, data...)

for suitable fitted_params and data, then

ScientificTypes.scitype(output) <: S

See also LearnAPI.predict_joint_input_scitype.

New model implementations

Implementation is optional. The fallback return value is Any.

LearnAPI.predict_joint_input_type — Function

LearnAPI.predict_joint_input_type(model)

Return an upper bound on the type of input data guaranteed to work with the predict_joint operation.

Specifically, if T is the value returned and typeof(data) <: S, then the following low-level calls are allowed

data2 = LearnAPI.reformat(model, LearnAPI.predict_joint, data...)
LearnAPI.predict_joint(model, fitted_params, data2...)

Here fitted_params are the learned parameters returned by an appropriate call to LearnAPI.fit.

See also LearnAPI.predict_joint_input_scitype.

New model implementations

Implementation is optional. The fallback return value is Union{}. Should not be overloaded if LearnAPI.predict_joint_input_scitype is overloaded.

LearnAPI.predict_joint_output_type — Function

LearnAPI.predict_joint_output_type(model)

Return an upper bound on the type of the output of the predict_joint operation.

Specifically, if T is the value returned, and if

output, report = LearnAPI.predict_joint(model, fitted_params, data...)

for suitable fitted_params and data, then

typeof(output) <: T

See also LearnAPI.predict_joint_input_type.

New model implementations

Implementation is optional. The fallback return value is Any.

LearnAPI.transform_input_scitype — Function

 LearnAPI.transform_input_scitype(model)

Return an upper bound on the scitype of input data guaranteed to work with the transform operation.

Specifically, if S is the value returned and ScientificTypes.scitype(data) <: S, then the following low-level calls are allowed

 data2 = LearnAPI.reformat(model, LearnAPI.transform, data...)
 LearnAPI.transform(model, fitted_params, data2...)

Here fitted_params are the learned parameters returned by an appropriate call to LearnAPI.fit.

See also LearnAPI.transform_input_type.

New model implementations

Implementation is optional. The fallback return value is Union{}. Should not be overloaded if LearnAPI.transform_input_type is overloaded.

LearnAPI.transform_output_scitype — Function

LearnAPI.transform_output_scitype(model)

Return an upper bound on the scitype of the output of the transform operation.

Specifically, if S is the value returned, and if

output, report = LearnAPI.transform(model, fitted_params, data...)

for suitable fitted_params and data, then

ScientificTypes.scitype(output) <: S

See also LearnAPI.transform_input_scitype.

New model implementations

Implementation is optional. The fallback return value is Any.

LearnAPI.transform_input_type — Function

LearnAPI.transform_input_type(model)

Return an upper bound on the type of input data guaranteed to work with the transform operation.

Specifically, if T is the value returned and typeof(data) <: S, then the following low-level calls are allowed

data2 = LearnAPI.reformat(model, LearnAPI.transform, data...)
LearnAPI.transform(model, fitted_params, data2...)

Here fitted_params are the learned parameters returned by an appropriate call to LearnAPI.fit.

See also LearnAPI.transform_input_scitype.

New model implementations

Implementation is optional. The fallback return value is Union{}. Should not be overloaded if LearnAPI.transform_input_scitype is overloaded.

LearnAPI.transform_output_type — Function

LearnAPI.transform_output_type(model)

Return an upper bound on the type of the output of the transform operation.

Specifically, if T is the value returned, and if

output, report = LearnAPI.transform(model, fitted_params, data...)

for suitable fitted_params and data, then

typeof(output) <: T

See also LearnAPI.transform_input_type.

New model implementations

Implementation is optional. The fallback return value is Any.

LearnAPI.inverse_transform_input_scitype — Function

 LearnAPI.inverse_transform_input_scitype(model)

Return an upper bound on the scitype of input data guaranteed to work with the inverse_transform operation.

Specifically, if S is the value returned and ScientificTypes.scitype(data) <: S, then the following low-level calls are allowed

 data2 = LearnAPI.reformat(model, LearnAPI.inverse_transform, data...)
 LearnAPI.inverse_transform(model, fitted_params, data2...)

Here fitted_params are the learned parameters returned by an appropriate call to LearnAPI.fit.

See also LearnAPI.inverse_transform_input_type.

New model implementations

Implementation is optional. The fallback return value is Union{}. Should not be overloaded if LearnAPI.inverse_transform_input_type is overloaded.

LearnAPI.inverse_transform_output_scitype — Function

LearnAPI.inverse_transform_output_scitype(model)

Return an upper bound on the scitype of the output of the inverse_transform operation.

Specifically, if S is the value returned, and if

output, report = LearnAPI.inverse_transform(model, fitted_params, data...)

for suitable fitted_params and data, then

ScientificTypes.scitype(output) <: S

See also LearnAPI.inverse_transform_input_scitype.

New model implementations

Implementation is optional. The fallback return value is Any.

LearnAPI.inverse_transform_input_type — Function

LearnAPI.inverse_transform_input_type(model)

Return an upper bound on the type of input data guaranteed to work with the inverse_transform operation.

Specifically, if T is the value returned and typeof(data) <: S, then the following low-level calls are allowed

data2 = LearnAPI.reformat(model, LearnAPI.inverse_transform, data...)
LearnAPI.inverse_transform(model, fitted_params, data2...)

Here fitted_params are the learned parameters returned by an appropriate call to LearnAPI.fit.

See also LearnAPI.inverse_transform_input_scitype.

New model implementations

Implementation is optional. The fallback return value is Union{}. Should not be overloaded if LearnAPI.inverse_transform_input_scitype is overloaded.

LearnAPI.inverse_transform_output_type — Function

LearnAPI.inverse_transform_output_type(model)

Return an upper bound on the type of the output of the inverse_transform operation.

Specifically, if T is the value returned, and if

output, report = LearnAPI.inverse_transform(model, fitted_params, data...)

for suitable fitted_params and data, then

typeof(output) <: T

See also LearnAPI.inverse_transform_input_type.

New model implementations

Implementation is optional. The fallback return value is Any.