Predict and Other Operations

Summary Methods like predict and transform, that generally depend on learned parameters, are called operations. All implemented operations must be included in the output of the methods model trait. When an operation returns a target proxy, it must make a target_proxy declaration.

An operation is any method with signature some_operation(model, fitted_params, data...). Here fitted_params is the learned parameters object, as returned by LearnAPI.fit(model, ...), which will be nothing if fit is not implemented (true for models that do not generalize to new data). For example, LearnAPI.predict in the following code snippet is an operation:

fitted_params, state, fit_report = LearnAPI.fit(some_model, 1, X, y)
ŷ, predict_report = LearnAPI.predict(some_model, fitted_params, Xnew)

method	compulsory?	fallback	requires

LearnAPI.predict	no	none

LearnAPI.predict_mode	no	none	predict

LearnAPI.predict_mean	no	none	predict

LearnAPI.predict_median	no	none	predict

LearnAPI.predict_joint	no	none

LearnAPI.transform	no	none

LearnAPI.inverse_transform	no	none	transform

† MLJ only. MLJBase provides fallbacks for predict_mode, predict_mean and predict_median by broadcasting methods from Statistics and StatsBase over the results of predict.

General requirements

Only implement predict_joint for outputing a single multivariate probability distribution for multiple target predictions, as described further at LearnAPI.predict_joint.

Each operation explicitly implemented or overloaded must be included in the return value of LearnAPI.methods.

Predict or transform?

If the model has a target, as defined under Scope and undefined notions, then only predict or predict_joint can be used to generate a corresponding target proxy.

If an operation is to have an inverse operation, then it cannot be predict - use transform and inverse_transform.

If only a single operation is implemented, and there is no target variable, use transform.

Here an "inverse" of transform is very broadly understood as any operation that can be applied to the output of transform to obtain an object of the same form as the input of transform; for example this includes one-sided inverses, and approximate one-sided inverses.

Target proxies

In the case that a model has the concept of a target variable, as described under Scope and undefined notions, the output of predict may have the form of a proxy for the target, such as a vector of truth-probabilities for binary targets.

We assume the reader is already familiar with the notion of a target variable in supervised learning, but target variables are not limited to supervised models. For example, we may regard the "outlier"/"inlier" assignments in unsupervised anomaly detection as a target. A target proxy in this example would be probabilities for outlierness, as these can be paired with "outlier"/"inlier" labels assigned by humans, using, say, area under the ROC curve, to quantify performance.

Similarly, the integer labels assigned to some observations by a clustering algorithm can be regarded as a target variable. The labels obtained can be paired with human labels using, say, the Rand index.

The kind of proxy one has is informally classified by a subtype of the abstract type LearnAPI.TargetProxy. These types are intended for dispatch outside of LearnAPI.jl and have no fields.

type	form of an observation

LearnAPI.TrueTarget	same as target observations (possible requirement: observations have same type as target observations)

LearnAPI.Sampleable	object that can be sampled to obtain object of the same form as target observation (possible requirement: observation implements Base.rand)

LearnAPI.Distribution	explicit probability density/mass function whose sample space is all possible target observations (possible requirement: observation implements Distributions.pdf and Base.rand)

LearnAPI.LogDistribution	explicit log-probability density/mass function whose sample space is possible target observations (possible requirement: observation implements Distributions.logpdf and Base.rand)

† LearnAPI.Probability	raw numerical probability or probability vector

† LearnAPI.LogProbability	log-probability or log-probability vector

† LearnAPI.Parametric	a list of parameters (e.g., mean and variance) describing some distribution

LearnAPI.LabelAmbiguous	same form as the (multi-class) target, but selected from new unmatched labels of possibly unequal number (as in, e.g., clustering)

LearnAPI.LabelAmbiguousSampleable	sampleable version of LabelAmbiguous; see Sampleable above

LearnAPI.LabelAmbiguousDistribution	pdf/pmf version of LabelAmbiguous; see Distribution above

LearnAPI.ConfidenceInterval	confidence interval (possible requirement: observation isa Tuple{Real,Real})

LearnAPI.Set	finite but possibly varying number of target observations (possible requirement: observation isa Set{target_observation_type})

LearnAPI.ProbabilisticSet	as for Set but labelled with probabilities (not necessarily summing to one)

LearnAPI.SurvivalFunction	survival function (possible requirement: observation is single-argument function mapping Real to Real)

LearnAPI.SurvivalDistribution	probability distribution for survival time (possible requirement: observation have type Distributions.ContinuousUnivariateDistribution)

† MLJ only. To avoid ambiguities in representation, these options are disallowed, in favour of preceding alternatives.

Warning

The "possible requirement"s listed are not part of LearnAPI.jl.

An operation with target proxy as output must declare a TargetProxy instance using the LearnAPI.target_proxy, as in

LearnAPI.target_proxy(::Type{<:SomeModel}) = (predict=LearnAPI.Distribution(),)

If predict_joint is implemented, then a target_proxy declaration is also required, but the interpretation is slightly different. This is because the output of predict_joint is not a number of observations but a single object. The trait value should be an instance of one of the following types:

type	form of output of predict_joint(model, _, data)

LearnAPI.Sampleable	object that can be sampled to obtain a vector whose elements have the form of target observations; the vector length matches the number of observations in data.

LearnAPI.Distribution	explicit probability density/mass function whose sample space is vectors of target observations; the vector length matches the number of observations in data

LearnAPI.LogDistribution	explicit log-probability density/mass function whose sample space is vectors of target observations; the vector length matches the number of observations in data

See more at LearnAPI.predict_joint below.

Operation-specific details

LearnAPI.predict — Function

LearnAPI.predict(model, fitted_params, data...)

Return (ŷ, report) where ŷ are the predictions, or prediction-like output (such as probabilities), for a machine learning model model, with learned parameters fitted_params, as returned by a preceding call to LearnAPI.fit(model, ...). Here report contains ancilliary byproducts of the computation, or is nothing; data is a tuple of data objects, generally a single object representing new observations not seen in training.

New model implementations

If implemented, include :predict in the tuple returned by the LearnAPI.methods trait.

If performance_measureable = true, then ŷ must be:

either an array or table with the same number of observations as each element of data; it cannot be a lazy object, such as a DataLoader

target-like; see LearnAPI.paradigm for specifics.

Otherwise there are no restrictions on what predict may return, apart from what the implementation itself promises, by making an optional LearnAPI.output_scitypes declaration.

If predict is computing a target proxy, as defined in the MLJLearn documentation, then a LearnAPI.target_proxy declaration is required, as in

LearnAPI.target_proxy(::Type{<:SomeModel}) = (predict=LearnAPI.Distribution,)

Do LearnAPI.target_proxy() to list the available kinds.

By default, it is expected that data has length one. Otherwise, LearnAPI.input_scitypes must be overloaded.

See also LearnAPI.fit, LearnAPI.predict_mean, LearnAPI.predict_mode, LearnAPI.predict_median.

LearnAPI.predict_mean — Function

LearnAPI.predict_mean(model, fitted_params, data...)

Same as LearnAPI.predict except replaces probababilistic predictions with mean values.

New model implementations

A fallback broadcasts mean over the first return value ŷ of LearnAPI.predict. An algorithm that computes probabilistic predictions may already need to predict mean values, and so overloading this method might enable a performance boost.

If overloaded, include :predict_mean in the tuple returned by the LearnAPI.methods trait.

See also LearnAPI.predict, LearnAPI.fit.

LearnAPI.predict_median — Function

LearnAPI.predict_median(model, fitted_params, data...)

Same as LearnAPI.predict except replaces probababilistic predictions with median values.

New model implementations

A fallback broadcasts median over the first return value ŷ of LearnAPI.predict. An algorithm that computes probabilistic predictions may already need to predict mean values, and so overloading this method might enable a performance boost.

If overloaded, include :predict_median in the tuple returned by the LearnAPI.methods trait.

See also LearnAPI.predict, LearnAPI.fit.

LearnAPI.predict_joint — Function

LearnAPI.predict_joint(model, fitted_params, data...)

For a supervised learning model, return (d, report), where d is some representation of the single probability distribution for the sample space $Y^n$. Here $Y$ is the space in which the target variable associated with model takes its values, and n is the number of observations in data. The specific form of the representation is given by LearnAPI.target_proxy(model).

Here fitted_params are the model's learned parameters, as returned by a preceding call to LearnAPI.fit. Here report contains ancilliary byproducts of the computation, or is nothing; data is a tuple of data objects, generally a single object representing new observations not seen in training. .

While the interpretation of this distribution depends on the model, marginalizing component-wise will generally deliver correlated univariate distributions, and these will generally not agree with those returned by LearnAPI.predict, if also implemented.

New model implementations

Only implement this method if model has an associated concept of target variable, as defined in the LearnAPI.jl documentation. A trait declaration LearnAPI.target_proxy, such as

LearnAPI.target_proxy(::Type{SomeModel}) = (; predict_joint=Sampleable())

is required. Here the possible kinds of target proxies are LearnAPI.Sampleable, LearnAPI.Distribution, and LearnAPI.LogDistribution.

If implemented, include :predict_joint in the tuple returned by the LearnAPI.methods trait. .

See also LearnAPI.fit, LearnAPI.predict.

LearnAPI.transform — Function

LearnAPI.transform(model, fitted_params, data...)

Return (output, report), where output is some kind of transformation of data, provided by model, based on the learned parameters fitted_params, as returned by a preceding call to LearnAPI.fit(model, ...) (which could be nothing for models that do not generalize to new data, such as "static transformers"). Here report contains ancilliary byproducts of the computation, or is nothing; data is a tuple of data objects, generally a single object representing new observations not seen in training.

New model implementations

If implemented, include :transform in the tuple returned by the LearnAPI.methods trait.

See also LearnAPI.inverse_transform, LearnAPI.fit, LearnAPI.predict,

LearnAPI.inverse_transform — Function

LearnAPI.inverse_transform(model, fitted_params, data)

Return (data_inverted, report), where data_inverted is valid input to the call

LearnAPI.transform(model, fitted_params, data_inverted)

Here report contains ancilliary byproducts of the computation, or is nothing; data is a tuple of data objects, generally a single object representing new observations not seen in training.

Typically, the map

data -> first(inverse_transform(model, fitted_params, data))

will be an inverse, approximate inverse, right inverse, or approximate right inverse, for the map

data -> first(transform(model, fitted_params, data))

For example, if transform corresponds to a projection, inverse_transform is the corresponding embedding.

New model implementations

If implemented, include :transform in the tuple returned by the LearnAPI.methods trait.

See also LearnAPI.fit, LearnAPI.predict,