Optional Data Interface

Summary. Implement getobs to articulate how to generate individual observations from data consumed by a LearnAPI algorithm. Implement reformat to provide a higher level interface the means to avoid repeating transformations from user representations of data (such as a dataframe) and algorithm-specific representations (such as a matrix).

Resampling

To aid in programmatic resampling, such as cross-validation, it is helpful if each machine learning algorithm articulates how the data it consumes can be subsampled - that is, how a subset of observations can be extracted from that data. Another advantage of doing so is to mitigate some of the ambiguities around structuring observations within the container: Are the observations in a matrix the rows or the columns?

In LearnAPI, an implementation can articulate a subsampling method by implementing LearnAPI.getobs(algorithm, func, I, data...) for each function func consuming data, such as fit and predict. Examples are given below.

LearnAPI.getobs — Function

LearnAPI.getobs(algorithm, LearnAPI.fit, I, data...)

Return a subsample of data consisting of all observations with indices in I. Here data is data of the form expected in a call like LearnAPI.fit(algorithm, verbosity, data...; metadata...).

Always returns a tuple of the same length as data.

LearnAPI.getobs(algorithm, operation, I, data...)

Return a subsample of data consisting of all observations with indices in I. Here data is data of the form expected in a call of the specified operation, e.g., in a call like LearnAPI.predict(algorithm, data...), if operation = LearnAPI.predict. Possible values for operation are: LearnAPI.predict, LearnAPI.transform, LearnAPI.inverse_transform.

Always returns a tuple of the same length as data.

New implementations

Implementation is optional. If implemented, then ordinarily implemented for each signature of fit and operation implemented for algorithm.

If implemented, you must include :reformat in the tuple returned by the LearnAPI.functions trait.

The subsample returned must be acceptable in place of data in calls of the function named in the second argument.

Sample implementation

Suppose that MyClassifier is an algorithm type for simple supervised classification, with LearnAPI.fit(algorithm::MyClassifier, verbosity, A, y) and predict(algorithm::MyClassifier, fitted_params, A) implemented assuming the target y is an ordinary abstract vector and the features A is an abstract matrix with columns as observations. Then the following is a valid implementation of getobs:

LearnAPI.getobs(::MyClassifier, ::typeof(LearnAPI.fit), I, A, y) =
    (view(A, :, I), view(y, I))
LearnAPI.getobs(::MyClassifier, ::typeof(LearnAPI.predict), I, A) = (view(A, :, I),)

Preprocessing

So that a higher level interface can avoid unnecessarily repeating calls to convert user-supplied data (e.g., a dataframe) into some performant, algorithm-specific representation, an algorithm can move such data conversions out of fit, predict, etc., and into an implementation of LearnAPI.reformat created for each signature of such methods that are implemented. Examples are given below.

LearnAPI.reformat — Function

LearnAPI.reformat(algorithm, LearnAPI.fit, user_data...; metadata...)

Return the algorithm-specific representations (data, metadata) of user-supplied (user_data, user_metadata), for consumption, after splatting, by LearnAPI.fit, LearnAPI.update! or LearnAPI.ingest!.

LearnAPI.reformat(algorithm, operation, user_data...)

Return the algorithm-specific representation data of user-supplied user_data, for consumption, after splatting, by the specified operation, dispatched on algorithm. Here operation is one of: LearnAPI.predict, LearnAPI.transform, LearnAPI.inverse_transform.

The following sample workflow illustrates the use of both versions of reformatabove. The data objects X, y, and Xtest are the user-supplied versions of data.

data, metadata = LearnAPI.reformat(algorithm, LearnAPI.fit, X, y; class_weights=some_dictionary)
fitted_params, state, fit_report = LearnAPI.fit(algorithm, 0, data...; metadata...)

test_data = LearnAPI.reformat(algorithm, LearnAPI.predict, Xtest)
ŷ, predict_report = LearnAPI.predict(algorithm, fitted_params, test_data...)

New implementations

Implementation of reformat is optional. The fallback simply slurps the supplied data/metadata. You will want to implement for each fit or operation signature implemented for algorithm.

If overloaded, you must include :reformat in the tuple returned by the LearnAPI.functions trait.

Ideally, any potentially expensive transformation of user-supplied data that is carried out during training only once, at the beginning, should occur in reformat instead of fit/update!/ingest!.

Note that the first form of reformat, for operations, should always return a tuple, because the output is splat in calls to the operation (see the sample workflow above). Similarly, in the return value (data, metadata) for the fit variant, data is always a tuple and metadata always a named tuple (or Base.Pairs object). If there is no metadata, a NamedTuple() can be returned in its place.

Example implementation

Suppose that MyClassifier is an algorithm type for simple supervised classification, with LearnAPI.fit(algorithm::MyClassifier, verbosity, A, y; names=...) and predict(algorithm::MyClassifier, fitted_params, A) implemented assuming that the target y is an ordinary vector, the features A is a matrix with columns as observations, and names are the names of the features. Then, supposing users supply features in tabular form, but target as expected, then we can provide the following implementation of reformat:

using Tables
function LearnAPI.reformat(::MyClassifier, ::typeof(LearnAPI.fit), X, y)
    names = Tables.schema(Tables.rows(X)).names
    return ((Tables.matrix(X)', y), (; names))
end
LearnAPI.reformat(::MyClassifier, ::typeof(LearnAPI.predict), X) = (Tables.matrix(X)',)