Accessor Functions

Summary. While byproducts of training are ordinarily recorded in the report component of the output of fit/update!/ingest!, some families of models report an item that is likely shared by multiple model types, and it is useful to have common interface for accessing these directly. Training losses and feature importances are two examples.

LearnAPI.feature_importances — Function

LearnAPI.feature_importances(model, fitted_params, report)

Return the model-specific feature importances of model, given fitted_params and report, as returned by LearnAPI.fit, LearnAPI.update! or LearnAPI.ingest!. The value returned has the form of an abstract vector of feature::Symbol => importance::Real pairs (e.g [:gender => 0.23, :height => 0.7, :weight => 0.1]).

The model supports feature importances if :feature_importance in LearnAPI.functions(model).

If for some reason a model is sometimes unable to report feature importances, then feature_importances will return all importances as 0.0, as in [:gender => 0.0, :height => 0.0, :weight => 0.0].

New model implementations

LearnAPI.feature_importances(model::SomeModelType, fitted_params, report) may be overloaded for any type SomeModelType whose instances are models in the LearnAPI sense. If a model can report multiple feature importance types, then the specific type to be reported should be controlled by a hyperparameter (i.e., by some property of model).

If implemented, you must include :feature_importances in the tuple returned by the LearnAPI.functions trait. .

LearnAPI.training_losses — Function

training_losses(model, fitted_params, report)

Return the training losses for model, given fitted_params and report, as returned by LearnAPI.fit, LearnAPI.update! or LearnAPI.ingest!.

New model implementations

Implement for iterative models that compute and record training losses as part of training (e.g. neural networks).

If implemented, you must include :training_losses in the tuple returned by the LearnAPI.functions trait. .

LearnAPI.training_scores — Function

training_scores(model, fitted_params, report)

Return the training scores for model, given fitted_params and report, as returned by LearnAPI.fit, LearnAPI.update! or LearnAPI.ingest!.

New model implementations

Implement for models, such as outlier detection models, which associate a score with each observation during training, where these scores are of interest in later processes (e.g, in defining normalized scores on new data).

If implemented, you must include :training_scores in the tuple returned by the LearnAPI.functions trait. .

LearnAPI.training_labels — Function

training_labels(model, fitted_params, report)

Return the training labels for model, given fitted_params and report, as returned by LearnAPI.fit, LearnAPI.update! or LearnAPI.ingest!.

New model implementations

If implemented, you must include :training_labels in the tuple returned by the LearnAPI.functions trait. .