Fit, update! and ingest!

Summary. Models that learn, i.e., generalize to new data, must overload fit; the fallback performs no operation and returns all nothing. Implement update! if certain hyper-parameter changes do not necessitate retraining from scratch (e.g., increasing iteration parameters). Implement ingest! to implement incremental learning.

method	fallback	compulsory?	requires

LearnAPI.fit	does nothing, returns (nothing, nothing, nothing)	no

LearnAPI.update!	calls fit	no	LearnAPI.fit

LearnAPI.ingest!	none	no	LearnAPI.fit

All three methods above return a triple (fitted_params, state, report) whose components are explained under LearnAPI.fit below. Items that might be returned in report include: feature rankings/importances, SVM support vectors, clustering centres, methods for visualizing training outcomes, methods for saving learned parameters in a custom format, degrees of freedom, deviances. Precisely what report includes might be controlled by model hyperparameters, especially if there is a performance cost to it's inclusion.

Implement fit unless all operations, such as predict and transform, ignore their fitted_params argument (which will be nothing). This is the case for many models that have hyperparameters, but do not generalize to new data, such as a basic DBSCAN clustering algorithm.

The update! method is intended for all subsequent calls to train a model using the same observations, but with possibly altered hyperparameters (model argument). A fallback implementation simply calls fit. The main use cases for implementing update are: (i) warm-restarting iterative models, and (ii) "smart" training of composite models, such as linear pipelines. Here "smart" means that hyperparameter changes only trigger the retraining of downstream components.

The ingest! method supports incremental learning (same hyperparameters, but new training observations). Like update!, it depends on the output a preceding fit or ingest! call.

LearnAPI.fit — Function

LearnAPI.fit(model, verbosity, data...; metadata...)

Fit model to the provided data and metadata. With the exception of warnings, training should be silent if verbosity == 0. Lower values should suppress warnings; any integer ought to be admissible. Here:

model is a property-accessible object whose properties are the hyper-parameters of some machine learning algorithm; see also LearnAPI.ismodel.

data is a tuple of data objects with a common number of observations, for example, data = (X, y, w) where X is a table of features, y is a target vector with the same number of rows, and w a vector of per-observation weights.

metadata is for extra information pertaining to the data that is never iterated, for example, weights for target classes. Another example would be feature groupings in the group lasso algorithm.

Return value

Returns a tuple (fitted_params, state, report) where:

The fitted_params is the model's learned parameters (eg, the coefficients in a linear model) in a form understood by model operations. An operation is a method, like LearnAPI.predict or LearnAPI.transform, that has signature (model, fitted_params, data....); do LearnAPI.OPERATIONS to list. If some training outcome of user-interest is not needed for operations, it should be part of report instead (see below).

The state is for passing to LearnAPI.update! or LearnAPI.ingest!. For models that implement neither, state should be nothing.

The report records byproducts of training not in the fitted_params.

New model implementations

Overloading this method for new models is optional. A fallback performs no computation, returning (nothing, nothing, nothing).

Note that in LearnAPI.jl the word "data" is only defined informally, as an object generating "observations", which are not defined at all.

Note

The method is not permitted to mutate model. In particular, if model has a random number generator as a hyperparameter (property) then it must be copied before use.

If implemented, include :fit in the tuple returned by the LearnAPI.methods trait.

See also LearnAPI.update!, LearnAPI.ingest!.

LearnAPI.update! — Function

LearnAPI.update!(model, verbosity, fitted_params, state, data...; metadata...)d Based on the values of `state`, and `fitted_params` returned by a preceding call to

LearnAPI.fit, LearnAPI.ingest!, or LearnAPI.update!, update a model's learned parameters, returning new (or mutated) state and fitted_params.

Intended for retraining a model when the training data has not changed, but model properties (hyperparameters) may have changed. Specifically, the assumption is that data and metadata have the same values seen in the most recent call to fit/update!/ingest!.

The most common use case is for continuing the training of an iterative model: state is simply a copy of the model used in the last training call (fit, update! or ingest!) and this will include the current number of iterations as a property. If model and state differ only in the number of iterations (e.g., epochs in a neural network), which has increased, then the learned parameters (weights) are updated, rather computed ab initio. Otherwise, update! simply calls fit, to force retraining from scratch.

It is permitted to return mutated versions of state and fitted_params.

For incremental training (same model, new data) see instead LearnAPI.ingest!.

Return value

Same as LearnAPI.fit, namely a tuple (fitted_params, state, report). See LearnAPI.fit for details.

New model implementations

Overloading this method is optional. A fallback calls LearnAPIperforms.fit:

LearnAPI.update!(model, verbosity, fitted_params, state, data...; metadata...) =
    fit(model, verbosity, data; metadata...)

Note that in LearnAPI.jl the word "data" is only defined informally, as an object generating "observations", which are not defined at all.

Note

The method is not permitted to mutate model. In particular, if model has a random number generator as a hyperparameter (property) then it must be copied before use.

If implemented, include :fit in the tuple returned by the LearnAPI.methods trait.

See also LearnAPI.fit, LearnAPI.ingest!.

LearnAPI.ingest! — Function

LearnAPI.ingest!(model, verbosity, fitted_params, state, data...; metadata...)

For a model that supports incremental learning, update the learned parameters using data, which has typically not been seen before. The arguments state and fitted_params are the output of a preceding call to LearnAPI.fit, LearnAPI.ingest!, or LearnAPI.update!, of which mutated or new versions are returned.

For updating learned parameters using the same data but new hyperparameters, see instead LearnAPI.update!.

For incremental training, see instead `LearnAPI.ingest'.

Return value

Same as LearnAPI.fit, namely a tuple (fitted_params, state, report). See LearnAPI.fit for details.

New model implementations

Implementing this method is an optional. It has no fallback.

Note

The method is not permitted to mutate model. In particular, if model has a random number generator as a hyperparameter (property) then it must be copied before use.

If implemented, include :fit in the tuple returned by the LearnAPI.methods trait.

See also LearnAPI.fit, LearnAPI.update!.