Predict and Other Operations

Summary An method delivering output for some algorithm which has finished learning, applied to (new) data, is called an operation. The output depends on the fitted parameters associated with the algorithm, which is nothing for non-generalizing algorithms. Implement the predict operation when the output is predictions of a target variable or, more generally a proxy for the target, such as probability distributions. Otherwise implement transform and, optionally inverse_transform.

The methods predict, transform and inverse_transform are called operations. They are all dispatched on an algorithm, fitted parameters and data. The predict operation additionally includes a ::ProxyType argument in position two. If LearnAPI.fit is not implemented, then the fitted parameters will always be nothing.

Here's a snippet of code with a LearnAPI.predict call:

fitted_params, state, fit_report = LearnAPI.fit(some_algorithm, 1, X, y)
ŷ, predict_report = 
    LearnAPI.predict(some_algorithm, LearnAPI.TrueTarget(), fitted_params, Xnew)

method	compulsory?	fallback	requires

LearnAPI.predict	no	none

LearnAPI.transform	no	none

LearnAPI.inverse_transform	no	none	transform

General requirements

Operations always return a tuple (output, report) where output is the usual output (e.g., the target predictions if the operation is predict) and report includes byproducts of the computation, typically nothing unless the algorithm does not generalize to new data (does not implement fit).

If implementing a predict method, you must also make a LearnAPI.preferred_kind_of_proxy declaration.

The name of each operation explicitly overloaded must be included in the return value of the LearnAPI.functions trait.

Predict or transform?

If the algorithm has a notion of target variable, then implement a predict method for each supported kind of target proxy (TrueTarget(), Distribution(), etc). See Target proxies below.

If an operation is to have an inverse operation, then it cannot be predict - use transform, and (optionally) inverse_transform, for inversion, broadly understood. See LearnAPI.inverse_transform below.

Target proxies

The concept of target proxy is defined under Targets and target proxies. The available kinds of target proxy are classified by subtypes of LearnAPI.KindOfProxy. These types are intended for dispatch only and have no fields.

LearnAPI.KindOfProxy — Type

LearnAPI.KindOfProxy

Abstract type whose concrete subtypes T each represent a different kind of proxy for the target variable, associated with some algorithm. Instances T() are used to request the form of target predictions in LearnAPI.predict calls.

For example, LearnAPI.Distribution is a concrete subtype of LearnAPI.KindOfProxy and the call LearnAPI.predict(algorithm , LearnAPI.Distribution(), data...) returns a data object whose observations are probability density/mass functions, assuming algorithm supports predictions of that form.

Run subtypes(LearnAPI.KindOfProxy) and subtypes(LearnAPI.IID) to list all concrete subtypes of KindOfProxy.

Predictions for i.i.d models

LearnAPI.IID — Type

LearnAPI.IID <: LearnAPI.KindOfProxy

Abstract subtype of LearnAPI.KindOfProxy appropriate when requesting target predictions for statistical models where input-target pairs (x, y) are generated by an iid process.

More generally, if kind_of_proxy is an instance of LearnAPI.IID then, given data constisting of $n$ observations, the following must hold:

LearnAPI.predict(algorithm, kind_of_proxy, data...) == (ŷ, report) where ŷ is data also consisting of $n$ observations; and

The $j$th observation of ŷ, for any $j$, depends only on the $j$th observation of the provided data (no correlation between observations).

See also LearnAPI.KindOfProxy.

type	form of an observation

LearnAPI.None	has no declared relationship with a target variable

LearnAPI.TrueTarget	same as target observations

LearnAPI.Sampleable	object that can be sampled to obtain object of the same form as target observation

LearnAPI.Distribution	explicit probability density/mass function whose sample space is all possible target observations

LearnAPI.LogDistribution	explicit log-probability density/mass function whose sample space is possible target observations

† LearnAPI.Probability	raw numerical probability or probability vector

† LearnAPI.LogProbability	log-probability or log-probability vector

† LearnAPI.Parametric	a list of parameters (e.g., mean and variance) describing some distribution

LearnAPI.LabelAmbiguous	collections of labels (in case of multi-class target) but without a known correspondence to the original target labels (and of possibly different number) as in, e.g., clustering

LearnAPI.LabelAmbiguousSampleable	sampleable version of LabelAmbiguous; see Sampleable above

LearnAPI.LabelAmbiguousDistribution	pdf/pmf version of LabelAmbiguous; see Distribution above

LearnAPI.ConfidenceInterval	confidence interval (possible requirement: observation isa Tuple{Real,Real})

LearnAPI.Set	finite but possibly varying number of target observations

LearnAPI.ProbabilisticSet	as for Set but labeled with probabilities (not necessarily summing to one)

LearnAPI.SurvivalFunction	survival function (possible requirement: observation is single-argument function mapping Real to Real)

LearnAPI.SurvivalDistribution	probability distribution for survival time

LearnAPI.OutlierScore	numerical score reflecting degree of outlierness (not necessarily normalized)

LearnAPI.Continuous	real-valued approximation/interpolation of a discrete-valued target, such as a count (e.g., number of phone calls)

† Provided for completeness but discouraged to avoid ambiguities in representation.

Table of concrete subtypes of LearnAPI.IID <: LearnAPI.KindOfProxy.

In the following table of subtypes T <: LearnAPI.KindOfProxy not falling under the IID umbrella, the first return value of predict(algorithm, ::T, fitted_params, data...) is not divided into individual observations, but represents a single probability distribution for the sample space $Y^n$, where $Y$ is the space the target variable takes its values, and n is the number of observations in data.

type T	form of output of predict(algorithm, ::T, fitted_params, data...)

LearnAPI.JointSampleable	object that can be sampled to obtain a vector whose elements have the form of target observations; the vector length matches the number of observations in data.

LearnAPI.JointDistribution	explicit probability density/mass function whose sample space is vectors of target observations; the vector length matches the number of observations in data

LearnAPI.JointLogDistribution	explicit log-probability density/mass function whose sample space is vectors of target observations; the vector length matches the number of observations in data

Table of LearnAPI.KindOfProxy subtypes not subtyping LearnAPI.IID

Reference

LearnAPI.predict — Function

LearnAPI.predict(algorithm, kind_of_proxy::LearnAPI.KindOfProxy, fitted_params, data...)

Return (ŷ, report) where ŷ is the predictions (a data object with target predictions as observations) or a proxy for these, for the specified algorithm having learned parameters fitted_params (first object returned by LearnAPI.fit(algorithm, ...)). The report contains ancilliary byproducts of the computation, or is nothing; data is a tuple of data objects, generally a single object representing new observations not seen in training.

Where available, use kind_of_proxy=TrueTarget() for ordinary target predictions, and kind_of_proxy=Distribution() for PDF/PMF predictions. Always available is kind_of_proxy=LearnAPI.preferredkindof_proxy(algorithm)`.

For a full list of target proxy types, run subtypes(LearnAPI.KindOfProxy) and subtypes(LearnAPI.IID).

New implementations

If implemented, you must include :predict in the tuple returned by the LearnAPI.functions trait.

If implementing LearnAPI.predict, then a LearnAPI.preferred_kind_of_proxy declaration is required, as in

LearnAPI.preferred_kind_of_proxy(::Type{<:SomeAlgorithm}) = LearnAPI.Distribution()

which has the shorthand

@trait SomeAlgorithm preferred_kind_of_proxy=LearnAPI.Distribution()

The value of this trait must be an instance T(), where T <: LearnAPI.KindOfProxy.

See also LearnAPI.fit.

LearnAPI.transform — Function

LearnAPI.transform(algorithm, fitted_params, data...)

Return (output, report), where output is some kind of transformation of data, provided by algorithm, based on the learned parameters fitted_params (the first object returned by LearnAPI.fit(algorithm, ...)). The fitted_params could be nothing, in the case of algorithms that do not generalize to new data. The report contains ancilliary byproducts of the computation, or is nothing; data is a tuple of data objects, generally a single object representing new observations not seen in training.

New implementations

If implemented, you must include :transform in the tuple returned by the LearnAPI.functions trait.

See also LearnAPI.inverse_transform, LearnAPI.fit, LearnAPI.predict,

LearnAPI.inverse_transform — Function

LearnAPI.inverse_transform(algorithm, fitted_params, data)

Return (data_inverted, report), where data_inverted is valid input to the call

LearnAPI.transform(algorithm, fitted_params, data_inverted)

The report contains ancilliary byproducts of the computation, or is nothing; data is a tuple of data objects, generally a single object representing new observations not seen in training.

Typically, the map

data -> first(inverse_transform(algorithm, fitted_params, data))

will be an inverse, approximate inverse, right inverse, or approximate right inverse, for the map

data -> first(transform(algorithm, fitted_params, data))

For example, if transform corresponds to a projection, inverse_transform might be the corresponding embedding.

New implementations

If implemented, you must include :transform in the tuple returned by the LearnAPI.functions trait.

See also LearnAPI.fit, LearnAPI.predict,