predict, transform and inverse_transform

predict(model, kind_of_proxy, data)
transform(model, data)
inverse_transform(model, data)

When a method expects a tuple form of argument, data = (X1, ..., Xn), then a slurping signature is also provided, as in transform(model, X1, ..., Xn).

Typical worklows

Train some supervised algorithm:

model = fit(algorithm, (X, y)) # or `fit(algorithm, X, y)`

Predict probability distributions:

ŷ = predict(model, Distribution(), Xnew)

Generate point predictions:

ŷ = predict(model, Point(), Xnew)

Train a dimension-reducing algorithm:

model = fit(algorithm, X)
Xnew_reduced = transform(model, Xnew)

Apply an approximate right inverse:

inverse_transform(model, Xnew_reduced)

An advanced workflow

fitobs = obs(algorithm, (X, y)) # algorithm-specific repr. of data
model = fit(algorithm, MLUtils.getobs(fitobs, 1:100))
predictobs = obs(model, MLUtils.getobs(X, 101:150))
ŷ = predict(model, Point(), predictobs)

Implementation guide

method	compulsory?	fallback

predict	no	none

transform	no	none

inverse_transform	no	none

Predict or transform?

If the algorithm has a notion of target variable, then use predict to output each supported kind of target proxy (Point(), Distribution(), etc).

For output not associated with a target variable, implement transform instead, which does not dispatch on LearnAPI.KindOfProxy, but can be optionally paired with an implementation of inverse_transform, for returning (approximate) right inverses to transform.

One-liners combining fit and transform/predict

Algorithms may optionally overload transform to apply fit first, using the supplied data if required, and then immediately transform the same data. The same applies to predict. In that case the first argument of transform/predict is an algorithm instead of the output of fit:

predict(algorithm, kind_of_proxy, data) # `fit` implied
transform(algorithm, data) # `fit` implied

For example, if fit(algorithm, X) is defined, then predict(algorithm, X) will be shorthand for

model = fit(algorithm, X)
predict(model, X)

Reference

LearnAPI.predict — Function

predict(model, kind_of_proxy::LearnAPI.KindOfProxy, data)
predict(model, data)

The first signature returns target predictions, or proxies for target predictions, for input features data, according to some model returned by fit. Where supported, these are literally target predictions if kind_of_proxy = Point(), and probability density/mass functions if kind_of_proxy = Distribution(). List all options with LearnAPI.kinds_of_proxy(algorithm), where algorithm = LearnAPI.algorithm(model).

The shortcut predict(model, data) calls the first method with an algorithm-specific kind_of_proxy, namely the first element of LearnAPI.kinds_of_proxy(algorithm), which lists all supported target proxies.

The argument model is anything returned by a call of the form fit(algorithm, ...).

Example

In the following, algorithm is some supervised learning algorithm with training features X, training target y, and test features Xnew:

model = fit(algorithm, (X, y)) # or `fit(algorithm, X, y)`
predict(model, Point(), Xnew)

See also fit, transform, inverse_transform.

Extended help

If predict supports data in the form of a tuple data = (X1, ..., Xn), then a slurping signature is also provided, as in predict(model, X1, ..., Xn).

Note predict does not mutate any argument, except in the special case LearnAPI.is_static(algorithm) == true.

New implementations

If there is no notion of a "target" variable in the LearnAPI.jl sense, or you need an operation with an inverse, implement transform instead.

Implementation is optional. Only the first signature is implemented, but each kind_of_proxy that gets an implementation must be added to the list returned by LearnAPI.kinds_of_proxy.

If implemented, you must include :(LearnAPI.predict) in the tuple returned by the LearnAPI.functions trait.

If, additionally, minimize(model) is overloaded, then the following identity must hold:

predict(minimize(model), args...) = predict(model, args...)

If LearnAPI.is_static(algorithm) is true, then predict may mutate it's first argument, but not in a way that alters the result of a subsequent call to predict, transform or inverse_transform. See more at fit.

Assumptions about data

By default, it is assumed that data supports the LearnAPI.RandomAccess interface; this includes all matrices, with observations-as-columns, most tables, and tuples thereof). See LearnAPI.RandomAccess for details. If this is not the case then an implementation must either: (i) overload obs to articulate how provided data can be transformed into a form that does support LearnAPI.RandomAccess; or (ii) overload the trait LearnAPI.data_interface to specify a more relaxed data API. Refer to document strings for details.

LearnAPI.transform — Function

transform(model, data)

Return a transformation of some data, using some model, as returned by fit.

For data that consists of a tuple, a slurping version is also provided, i.e., you can do transform(model, X1, X2, X3) in place of transform(model, (X1, X2, X3)).

Example

Below, X and Xnew are data of the same form.

For an algorithm that generalizes to new data ("learns"):

model = fit(algorithm, X; verbosity=0)
transform(model, Xnew)

For a static (non-generalizing) transformer:

model = fit(algorithm)
W = transform(model, X)

or, in one step (where supported):

W = transform(algorithm, X)

Note transform does not mutate any argument, except in the special case LearnAPI.is_static(algorithm) == true.

See also fit, predict, inverse_transform.

Extended help

New implementations

Implementation for new LearnAPI.jl algorithms is optional. A fallback provides the slurping version. If implemented, you must include :(LearnAPI.transform) in the tuple returned by the LearnAPI.functions trait.

If, additionally, minimize(model) is overloaded, then the following identity must hold:

transform(minimize(model), args...) = transform(model, args...)

If LearnAPI.is_static(algorithm) is true, then transform may mutate it's first argument, but not in a way that alters the result of a subsequent call to predict, transform or inverse_transform. See more at fit.

Assumptions about data

By default, it is assumed that data supports the LearnAPI.RandomAccess interface; this includes all matrices, with observations-as-columns, most tables, and tuples thereof). See LearnAPI.RandomAccess for details. If this is not the case then an implementation must either: (i) overload obs to articulate how provided data can be transformed into a form that does support LearnAPI.RandomAccess; or (ii) overload the trait LearnAPI.data_interface to specify a more relaxed data API. Refer to document strings for details.

LearnAPI.inverse_transform — Function

inverse_transform(model, data)

Inverse transform data according to some model returned by fit. Here "inverse" is to be understood broadly, e.g, an approximate right inverse for transform.

Example

In the following, algorithm is some dimension-reducing algorithm that generalizes to new data (such as PCA); Xtrain is the training input and Xnew the input to be reduced:

model = fit(algorithm, Xtrain)
W = transform(model, Xnew)       # reduced version of `Xnew`
Ŵ = inverse_transform(model, W)  # embedding of `W` in original space

See also fit, transform, predict.

Extended help

New implementations

Implementation is optional. If implemented, you must include :(LearnAPI.inverse_transform) in the tuple returned by the LearnAPI.functions trait.

If, additionally, minimize(model) is overloaded, then the following identity must hold:

inverse_transform(minimize(model), args...) = inverse_transform(model, args...)