target, weights, and features

Methods for extracting parts of training data:

LearnAPI.target(learner, data) -> <target variable>
LearnAPI.weights(learner, data) -> <per-observation weights>
LearnAPI.features(learner, data) -> <training "features", suitable input for `predict` or `transform`>

Here data is something supported in a call of the form fit(learner, data).

Typical workflow

Not typically appearing in a general user's workflow but useful in meta-alagorithms, such as cross-validation (see the example in obs and Data Interfaces).

Supposing learner is a supervised classifier predicting a one-dimensional vector target:

model = fit(learner, data)
X = LearnAPI.features(learner, data)
y = LearnAPI.target(learner, data)
ŷ = predict(model, Point(), X)
training_loss = sum(ŷ .!= y)

Implementation guide

method	fallback	compulsory?

LearnAPI.target	returns nothing	no

LearnAPI.weights	returns nothing	no

LearnAPI.features	see docstring	if fallback insufficient

Reference

LearnAPI.target — Function

LearnAPI.target(learner, data) -> target

Return, for each form of data supported in a call of the form fit(learner, data), the target variable part of data. If nothing is returned, the learner does not see a target variable in training (is unsupervised).

Refer to LearnAPI.jl documentation for the precise meaning of "target".

New implementations

A fallback returns nothing. Must be implemented if fit consumes data including a target variable.

If overloaded, you must include :(LearnAPI.target) in the tuple returned by the LearnAPI.functions trait.

LearnAPI.weights — Function

LearnAPI.weights(learner, data) -> weights

Return, for each form of data supported in a call of the form fit(learner, data), the per-observation weights part of data. Where nothing is returned, no weights are part of data, which is to be interpreted as uniform weighting.

New implementations

Overloading is optional. A fallback returns nothing.

If overloaded, you must include :(LearnAPI.weights) in the tuple returned by the LearnAPI.functions trait.

LearnAPI.features — Function

LearnAPI.features(learner, data)

Return, for each form of data supported in a call of the form fit(learner, data), the "features" part of data (as opposed to the target variable, for example).

The returned object X may always be passed to predict or transform, where implemented, as in the following sample workflow:

model = fit(learner, data)
X = features(data)
ŷ = predict(learner, kind_of_proxy, X) # eg, `kind_of_proxy = Point()`

The returned object has the same number of observations as data. For supervised models (i.e., where :(LearnAPI.target) in LearnAPI.functions(learner)) ŷ above is generally intended to be an approximate proxy for LearnAPI.target(learner, data), the training target.

New implementations

That the output can be passed to predict and/or transform, and has the same number of observations as data, are the only contracts. A fallback returns first(data) if data is a tuple, and otherwise returns data.

Overloading may be necessary if obs(learner, data) is overloaded to return some learner-specific representation of training data. For density estimators, whose fit typically consumes only a target variable, you should overload this method to return nothing.