Algorithm Traits

Algorithm traits are simply functions whose sole argument is an algorithm.

Traits promise specific algorithm behavior, such as: This algorithm can make point or probabilistic predictions or This algorithm is supervised (sees a target in training). They may also record more mundane information, such as a package license.

Trait summary

Overloadable traits

In the examples column of the table below, Continuous is a name owned the package ScientificTypesBase.jl.

trait	return value	fallback value	example

LearnAPI.constructor(algorithm)	constructor for generating new or modified versions of algorithm	(no fallback)	RidgeRegressor

LearnAPI.functions(algorithm)	functions you can apply to algorithm or associated model (traits excluded)	()	(:fit, :predict, :minimize, :(LearnAPI.algorithm), :obs)

LearnAPI.kinds_of_proxy(algorithm)	instances kind of KindOfProxy for which an implementation of LearnAPI.predict(algorithm, kind, ...) is guaranteed.	()	(Distribution(), Interval())

LearnAPI.tags(algorithm)	lists one or more suggestive algorithm tags from LearnAPI.tags()	()	(:regression, :probabilistic)

LearnAPI.is_pure_julia(algorithm)	true if implementation is 100% Julia code	false	true

LearnAPI.pkg_name(algorithm)	name of package providing core code (may be different from package providing LearnAPI.jl implementation)	"unknown"	"DecisionTree"

LearnAPI.pkg_license(algorithm)	name of license of package providing core code	"unknown"	"MIT"

LearnAPI.doc_url(algorithm)	url providing documentation of the core code	"unknown"	"https://en.wikipedia.org/wiki/Decision_tree_learning"

LearnAPI.load_path(algorithm)	string locating name returned by LearnAPI.constructor(algorithm), beginning with a package name	"unknown"`	FastTrees.LearnAPI.DecisionTreeClassifier

LearnAPI.is_composite(algorithm)	true if one or more properties of algorithm may be an algorithm	false	true

LearnAPI.human_name(algorithm)	human name for the algorithm; should be a noun	type name with spaces	"elastic net regressor"

LearnAPI.iteration_parameter(algorithm)	symbolic name of an iteration parameter	nothing	:epochs

LearnAPI.data_interface(algorithm)	Interface implemented by objects returned by obs	Base.HasLength() (supports MLUtils.getobs/numobs)	Base.SizeUnknown() (supports iterate)

LearnAPI.fit_observation_scitype(algorithm)	upper bound on scitype(observation) for observation in data ensuring fit(algorithm, data) works	Union{}	Tuple{AbstractVector{Continuous}, Continuous}

LearnAPI.target_observation_scitype(algorithm)	upper bound on the scitype of each observation of the targget	Any	Continuous

LearnAPI.is_static(algorithm)	true if fit consumes no data	false	true

Derived Traits

The following are provided for convenience but should not be overloaded by new algorithms:

trait	return value	example

LearnAPI.name(algorithm)	algorithm type name as string	"PCA"

LearnAPI.is_algorithm(algorithm)	true if algorithm is LearnAPI.jl-compliant	true

LearnAPI.target(algorithm)	true if fit sees a target variable; see LearnAPI.target	false

LearnAPI.weights(algorithm)	true if fit supports per-observation; see LearnAPI.weights	false

Implementation guide

A single-argument trait is declared following this pattern:

LearnAPI.is_pure_julia(algorithm::MyAlgorithmType) = true

A shorthand for single-argument traits is available:

@trait MyAlgorithmType is_pure_julia=true

Multiple traits can be declared like this:

@trait(
    MyAlgorithmType,
    is_pure_julia = true,
    pkg_name = "MyPackage",
)

The global trait contract

To ensure that trait metadata can be stored in an external algorithm registry, LearnAPI.jl requires:

Finiteness: The value of a trait is the same for all algorithms with same value of LearnAPI.constructor(algorithm). This typically means trait values do not depend on type parameters! If is_composite(algorithm) = true, this requirement is dropped.

Low level deserializability: It should be possible to evaluate the trait value when LearnAPI is the only imported module.

Because of 1, combining a lot of functionality into one algorithm (e.g. the algorithm can perform both classification or regression) can mean traits are necessarily less informative (as in LearnAPI.target_observation_scitype(algorithm) = Any).

Reference

LearnAPI.constructor — Function

Learn.API.constructor(algorithm)

Return a keyword constructor that can be used to clone algorithm:

julia> algorithm.lambda
0.1
julia> C = LearnAPI.constructor(algorithm)
julia> algorithm2 = C(lambda=0.2)
julia> algorithm2.lambda
0.2

New implementations

All new implementations must overload this trait.

Attach public LearnAPI.jl-related documentation for an algorithm to the constructor, not the algorithm struct.

It must be possible to recover an algorithm from the constructor returned as follows:

properties = propertynames(algorithm)
named_properties = NamedTuple{properties}(getproperty.(Ref(algorithm), properties))
@assert algorithm == LearnAPI.constructor(algorithm)(; named_properties...)

which can be tested with @assert LearnAPI.clone(algorithm) == algorithm.

The keyword constructor provided by LearnAPI.constructor must provide default values for all properties, with the exception of those that can take other LearnAPI.jl algorithms as values. These can be provided with the default nothing, with the constructor throwing an error if the default value persists.

LearnAPI.functions — Function

LearnAPI.functions(algorithm)

Return a tuple of expressions representing functions that can be meaningfully applied with algorithm, or an associated model (object returned by fit(algorithm, ...), as the first argument. Algorithm traits (methods for which algorithm is the only argument) are excluded.

The returned tuple may include expressions like :(DecisionTree.print_tree), which reference functions not owned by LearnAPI.jl.

The understanding is that algorithm is a LearnAPI-compliant object whenever the return value is non-empty.

Extended help

New implementations

All new implementations must overload this trait. Here's a checklist for elements in the return value:

expression	implementation compulsory?	include in returned tuple?

:(LearnAPI.fit)	yes	yes

:(LearnAPI.algorithm)	yes	yes

:(LearnAPI.minimize)	no	yes

:(LearnAPI.obs)	no	yes

:(LearnAPI.features)	no	yes, unless fit consumes no data

:(LearnAPI.target)	no	only if implemented

:(LearnAPI.weights)	no	only if implemented

:(LearnAPI.update)	no	only if implemented

:(LearnAPI.update_observations)	no	only if implemented

:(LearnAPI.update_features)	no	only if implemented

:(LearnAPI.predict)	no	only if implemented

:(LearnAPI.transform)	no	only if implemented

:(LearnAPI.inverse_transform)	no	only if implemented

<accessor functions>	no	only if implemented

Also include any implemented accessor functions, both those owned by LearnaAPI.jl, and any algorithm-specific ones. The LearnAPI.jl accessor functions are: LearnAPI.extras, LearnAPI.algorithm, LearnAPI.coefficients, LearnAPI.intercept, LearnAPI.tree, LearnAPI.trees, LearnAPI.feature_importances, LearnAPI.training_labels, LearnAPI.training_losses, LearnAPI.training_predictions, LearnAPI.training_scores and LearnAPI.components.

LearnAPI.kinds_of_proxy — Function

LearnAPI.kinds_of_proxy(algorithm)

Returns a tuple of all instances, kind, for which for which predict(algorithm, kind, data...) has a guaranteed implementation. Each such kind subtypes LearnAPI.KindOfProxy. Examples are Point() (for predicting actual target values) and Distributions() (for predicting probability mass/density functions).

The call predict(model, data) always returns predict(model, kind, data), where kind is the first element of the trait's return value.

See also LearnAPI.predict, LearnAPI.KindOfProxy.

Extended help

New implementations

Must be overloaded whenever predict is implemented.

Elements of the returned tuple must be instances of types in the return value of LearnAPI.kinds_of_proxy(), i.e., one of the following, described further in LearnAPI.jl documentation: ConfidenceInterval(), Continuous(), Distribution(), Expectile(), Fuzzy(), HazardFunction(), LabelAmbiguous(), LabelAmbiguousDistribution(), LabelAmbiguousFuzzy(), LabelAmbiguousSampleable(), LogDistribution(), LogProbability(), OutlierScore(), Parametric(), Point(), ProbabilisticFuzzy(), Probability(), Quantile(), Sampleable(), SurvivalDistribution(), SurvivalFunction(), SingleDistribution(), SingleLogDistribution(), SingleSampeable(), JointDistribution(), JointLogDistribution() and JointSampleable().

Suppose, for example, we have the following implementation of a supervised learner returning only probabilistic predictions:

LearnAPI.predict(algorithm::MyNewAlgorithmType, LearnAPI.Distribution(), Xnew) = ...

Then we can declare

@trait MyNewAlgorithmType kinds_of_proxy = (LearnaAPI.Distribution(),)

LearnAPI.jl provides the fallback for predict(model, data).

For more on target variables and target proxies, refer to the LearnAPI documentation.

LearnAPI.tags — Function

LearnAPI.tags(algorithm)

Lists one or more suggestive algorithm tags. Do LearnAPI.tags() to list all possible.

Warning

The value of this trait guarantees no particular behavior. The trait is intended for informal classification purposes only.

New implementations

This trait should return a tuple of strings, as in ("classifier", "text analysis").

LearnAPI.is_pure_julia — Function

LearnAPI.is_pure_julia(algorithm)

Returns true if training algorithm requires evaluation of pure Julia code only.

New implementations

The fallback is false.

LearnAPI.pkg_name — Function

LearnAPI.pkg_name(algorithm)

Return the name of the package module which supplies the core training algorithm for algorithm. This is not necessarily the package providing the LearnAPI interface.

Returns "unknown" if the algorithm implementation has not overloaded the trait.

New implementations

Must return a string, as in "DecisionTree".

LearnAPI.pkg_license — Function

LearnAPI.pkg_license(algorithm)

Return the name of the software license, such as "MIT", applying to the package where the core algorithm for algorithm is implemented.

LearnAPI.doc_url — Function

LearnAPI.doc_url(algorithm)

Return a url where the core algorithm for algorithm is documented.

Returns "unknown" if the algorithm implementation has not overloaded the trait.

New implementations

Must return a string, such as "https://en.wikipedia.org/wiki/Decision_tree_learning".

LearnAPI.load_path — Function

LearnAPI.load_path(algorithm)

Return a string indicating where in code the definition of the algorithm's constructor can be found, beginning with the name of the package module defining it. By "constructor" we mean the return value of LearnAPI.constructor(algorithm).

Implementation

For example, a return value of "FastTrees.LearnAPI.DecisionTreeClassifier" means the following julia code will not error:

import FastTrees
import LearnAPI
@assert FastTrees.LearnAPI.DecisionTreeClassifier == LearnAPI.constructor(algorithm)

Returns "unknown" if the algorithm implementation has not overloaded the trait.

LearnAPI.is_composite — Function

LearnAPI.is_composite(algorithm)

Returns true if one or more properties (fields) of algorithm may themselves be algorithms, and false otherwise.

See also LearnAPI.components.

New implementations

This trait should be overloaded if one or more properties (fields) of algorithm may take algorithm values. Fallback return value is false. The keyword constructor for such an algorithm need not prescribe defaults for algorithm-valued properties. Implementation of the accessor function LearnAPI.components is recommended.

The value of the trait must depend only on the type of algorithm.

LearnAPI.human_name — Function

LearnAPI.human_name(algorithm)

Return a human-readable string representation of typeof(algorithm). Primarily intended for auto-generation of documentation.

New implementations

Optional. A fallback takes the type name, inserts spaces and removes capitalization. For example, KNNRegressor becomes "knn regressor". Better would be to overload the trait to return "K-nearest neighbors regressor". Ideally, this is a "concrete" noun like "ridge regressor" rather than an "abstract" noun like "ridge regression".

LearnAPI.data_interface — Function

LearnAPI.data_interface(algorithm)

Return the data interface supported by algorithm for accessing individual observations in representations of input data returned by obs(algorithm, data) or obs(model, data), whenever algorithm == LearnAPI.algorithm(model). Here data is fit, predict, or transform-consumable data.

Possible return values are LearnAPI.RandomAccess, LearnAPI.FiniteIterable, and LearnAPI.Iterable.

See also obs.

New implementations

The fallback returns LearnAPI.RandomAccess, which applies to arrays, most tables, and tuples of these. See the doc-string for details.

LearnAPI.iteration_parameter — Function

LearnAPI.iteration_parameter(algorithm)

The name of the iteration parameter of algorithm, or nothing if the algorithm is not iterative.

New implementations

Implement if algorithm is iterative. Returns a symbol or nothing.

LearnAPI.fit_observation_scitype — Function

LearnAPI.fit_observation_scitype(algorithm)

Return an upper bound S on the scitype of individual observations guaranteed to work when calling fit: if observations = obs(algorithm, data) and ScientificTypes.scitype(o) <:S for each o in observations, then the call fit(algorithm, data) is supported.

Here, "for each o in observations" is understood in the sense of LearnAPI.data_interface(algorithm). For example, if LearnAPI.data_interface(algorithm) == Base.HasLength(), then this means "for o in MLUtils.eachobs(observations)".

See also LearnAPI.target_observation_scitype.

New implementations

Optional. The fallback return value is Union{}. Ordinarily, at most one of the following should be overloaded for given algorithm LearnAPI.fit_scitype, LearnAPI.fit_type, LearnAPI.fit_observation_scitype, LearnAPI.fit_observation_type.

LearnAPI.target_observation_scitype — Function

LearnAPI.target_observation_scitype(algorithm)

Return an upper bound S on the scitype of each observation of an applicable target variable. Specifically:

If :(LearnAPI.target) in LearnAPI.functions(algorithm) (i.e., fit consumes target variables) then "target" means anything returned by LearnAPI.target(algorithm, data), where data is an admissible argument in the call fit(algorithm, data).

S will always be an upper bound on the scitype of (point) observations that could be conceivably extracted from the output of predict.

To illustate the second case, suppose we have

model = fit(algorithm, data)
ŷ = predict(model, Sampleable(), data_new)

Then each individual sample generated by each "observation" of ŷ (a vector of sampleable objects, say) will be bound in scitype by S.

See also See also LearnAPI.fit_observation_scitype.

New implementations

Optional. The fallback return value is Any.

LearnAPI.is_static — Function

LearnAPI.is_static(algorithm)

Returns true if fit is called with no data arguments, as in fit(algorithm). That is, algorithm does not generalize to new data, and data is only provided at the predict or transform step.

For example, some clustering algorithms are applied with this workflow, to assign labels to the observations in X:

model = fit(algorithm) # no training data
labels = predict(model, X) # may mutate `model`!

# extract some byproducts of the clustering algorithm (e.g., outliers):
LearnAPI.extras(model)

New implementations

This trait, falling back to false, may only be overloaded when fit has no data arguments. See more at fit.